Noncompartmental Analysis for an Individual

It performs a noncompartmental analysis with one subject data. This will be deprecated. Use sNCA() instead.

IndiNCA(x, y, dose = 0, fit = "Linear", adm = "Extravascular", dur = 0,
               report = "Table", iAUC = "", uTime = "h", uConc = "ug/L", uDose = "mg")

Arguments

x
vector values of independent variable, usually time

y
vector values of dependent variable, usually concentration

dose
administered dose for the subject

fit
either of "Linear" or "Log" to indicate the way to calculate AUC and AUMC

adm
one of "Bolus" or "Infusion" or "Extravascular" to indicate drug administration mode

dur
infusion duration for constant infusion, otherwise 0

report
either of "Table" or "Text" to specify the type of return value

iAUC
data.frame with three columns, "Name", "Start", "End" to specify the invervals for partial (interval) AUC

uTime
unit of time

uConc
unit of concentration

uDose
unit of dose

Details

This performs a noncompartmental analysis for a subject. It returns practically the same result with the most popular commercial software.

Value

CMAX

maximum concentration, Cmax

CMAXD

dose normalized Cmax, CMAX / Dose, Cmax / Dose

TMAX

time of maximum concentration, Tmax

TLAG

time to observe the first non-zero concentration, for extravascular administration only

CLST

last positive concentration observed, Clast

CLSTP

last positive concentration predicted, Clast_pred

TLST

time of last positive concentration, Tlast

LAMZHL

half-life by lambda z, ln(2)/LAMZ

LAMZ

lambda_z negative of best fit terminal slope

LAMZLL

earliest time for LAMZ

LAMZUL

last time for LAMZ

LAMZNPT

number of points for LAMZ

CORRXY

correlation of log(concentration) and time

R2

R-squared

R2ADJ

R-squared adjusted

C0

back extrapolated concentration at time 0, for bolus intravascular administration only

AUCLST

AUC from 0 to TLST

AUCALL

AUC using all the given points, including trailing zero concentrations

AUCIFO

AUC infinity observed

AUCIFOD

AUCIFO / Dose

AUCIFP

AUC infinity predicted using CLSTP instead of CLST

AUCIFPD

AUCIFP / Dose

AUCPEO

AUC % extrapolation observed

AUCPEP

AUC % extrapolated for AUCIFP

AUCPBEO

AUC % back extrapolation observed, for bolus IV administration only

AUCPBEP

AUC % back extrapolation predicted with AUCIFP, for bolus IV administration only

AUMCLST

AUMC to the TLST

AUMCIFO

AUMC infinity observed using CLST

AUMCIFP

AUMC infinity determined by CLSTP

AUMCPEO

AUMC % extrapolated observed

AUMCPEP

AUMC % extrapolated predicted

MRTIVLST

mean residence time (MRT) to TLST, for intravascular administration

MRTIVIFO

mean residence time (MRT) infinity using CLST, for intravascular administration

MRTIVIFP

mean residence time (MRT) infinity using CLSTP, for intravascular administration

MRTEVLST

mean residence time (MRT) to TLST, for extravascular administration

MRTEVIFO

mean residence time (MRT) infinity using CLST, for extravascular administration

MRTEVIFP

mean residence time (MRT) infinity using CLSTP, for extravascular administration

VZO

volume of distribution determined by LAMZ and AUCIFO, for intravascular administration

VZP

volume of distribution determined by LAMZ and AUCIFP, for intravascular administration

VZFO

VZO for extravascular administration, VZO/F, F is bioavailability

VZFP

VZP for extravascular administration, VZP/F, F is bioavailability

CLO

clearance using AUCIFO, for intravascular administration

CLP

clearance using AUCIFP, for intravascular administration

CLFO

CLO for extravascular administration, CLO/F, F is bioavailability

CLFP

CLP for extravascular administration, CLP/F, F is bioavailability

VSSO

volume of distribution at steady state using CLST, for intravascular administration only

VSSP

volume of distribution at stead state using CLSTP, for intravascular administration only

References

Gabrielsson J, Weiner D. Pharmacokinetic and Pharmacodynamic Data Analysis - Concepts and Applications. 5th ed. 2016.

Shargel L, Yu A. Applied Biopharmaceutics and Pharmacokinetics. 7th ed. 2015.

Rowland M, Tozer TN. Clinical Pharmacokinetics and Pharmacodynamics - Concepts and Applications. 4th ed. 2011.

Gibaldi M, Perrier D. Pharmacokinetics. 2nd ed. revised and expanded. 1982.

See also

Examples

IndiNCA(Theoph[Theoph$Subject==1,"Time"], Theoph[Theoph$Subject==1, "conc"], dose=320, uConc="mg/L")

#>           b0         CMAX        CMAXD         TMAX         TLAG         CLST 
#>    2.3687851   10.5000000    0.0328125    1.1200000    0.0000000    3.2800000 
#>        CLSTP         TLST       LAMZHL         LAMZ       LAMZLL       LAMZUL 
#>    3.2801465   24.3700000   14.3043776    0.0484570    9.0500000   24.3700000 
#>      LAMZNPT       CORRXY           R2        R2ADJ           C0       AUCLST 
#>    3.0000000   -0.9999999    0.9999997    0.9999995           NA  148.9230500 
#>       AUCALL       AUCIFO      AUCIFOD       AUCIFP      AUCIFPD       AUCPEO 
#>  148.9230500  216.6119330    0.6769123  216.6149558    0.6769217   31.2489169 
#>       AUCPEP      AUCPBEO      AUCPBEP      AUMCLST      AUMCIFO      AUMCIFP 
#>   31.2498763           NA           NA 1459.0711035 4505.5348194 4505.6708646 
#>      AUMCPEO      AUMCPEP     MRTIVLST     MRTIVIFO     MRTIVIFP     MRTEVLST 
#>   67.6160287   67.6170065           NA           NA           NA    9.7974834 
#>     MRTEVIFO     MRTEVIFP          VZO          VZP         VZFO         VZFP 
#>   20.8000305   20.8003683           NA           NA   30.4867482   30.4863228 
#>          CLO          CLP         CLFO         CLFP         VSSO         VSSP 
#>           NA           NA    1.4772963    1.4772757           NA           NA 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Bolus", uConc="mg/L")

#>          b0        CMAX       CMAXD        TMAX        TLAG        CLST 
#> -1.72421060  1.50000000  0.06000000  0.25000000          NA  0.05000000 
#>       CLSTP        TLST      LAMZHL        LAMZ      LAMZLL      LAMZUL 
#>  0.05024851  8.00000000  4.37812701  0.15832048  5.00000000  8.00000000 
#>     LAMZNPT      CORRXY          R2       R2ADJ          C0      AUCLST 
#>  3.00000000 -0.99853229  0.99706673  0.99413345  2.39361702  2.04045213 
#>      AUCALL      AUCIFO     AUCIFOD      AUCIFP     AUCIFPD      AUCPEO 
#>  2.04045213  2.35626723  0.09425069  2.35783688  0.09431348 13.40319561 
#>      AUCPEP     AUCPBEO     AUCPBEP     AUMCLST     AUMCIFO     AUMCIFP 
#> 13.46084419 20.65564214 20.64189142  3.27125000  7.79255448  7.81502594 
#>     AUMCPEO     AUMCPEP    MRTIVLST    MRTIVIFO    MRTIVIFP    MRTEVLST 
#> 58.02082606 58.14153371  1.60319860  3.30716074  3.31448966          NA 
#>    MRTEVIFO    MRTEVIFP         VZO         VZP        VZFO        VZFP 
#>          NA          NA 67.01597804 66.97136466          NA          NA 
#>         CLO         CLP        CLFO        CLFP        VSSO        VSSP 
#> 10.61000197 10.60293876          NA          NA 35.08898193 35.14333088 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Infusion", dur=0.25, uConc="mg/L")

#>          b0        CMAX       CMAXD        TMAX        TLAG        CLST 
#> -1.72421060  1.50000000  0.06000000  0.25000000  0.00000000  0.05000000 
#>       CLSTP        TLST      LAMZHL        LAMZ      LAMZLL      LAMZUL 
#>  0.05024851  8.00000000  4.37812701  0.15832048  5.00000000  8.00000000 
#>     LAMZNPT      CORRXY          R2       R2ADJ          C0      AUCLST 
#>  3.00000000 -0.99853229  0.99706673  0.99413345          NA  1.74125000 
#>      AUCALL      AUCIFO     AUCIFOD      AUCIFP     AUCIFPD      AUCPEO 
#>  1.74125000  2.05706511  0.08228260  2.05863475  0.08234539 15.35270349 
#>      AUCPEP     AUCPBEO     AUCPBEP     AUMCLST     AUMCIFO     AUMCIFP 
#> 15.41724428          NA          NA  3.27125000  7.79255448  7.81502594 
#>     AUMCPEO     AUMCPEP    MRTIVLST    MRTIVIFO    MRTIVIFP    MRTEVLST 
#> 58.02082606 58.14153371  1.75367911  3.66319049  3.67121783          NA 
#>    MRTEVIFO    MRTEVIFP         VZO         VZP        VZFO        VZFP 
#>          NA          NA 76.76351746 76.70498779          NA          NA 
#>         CLO         CLP        CLFO        CLFP        VSSO        VSSP 
#> 12.15323712 12.14397067          NA          NA 44.51962266 44.58316170 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        

IndiNCA(Theoph[Theoph$Subject==1,"Time"], Theoph[Theoph$Subject==1, "conc"], dose=320,
        report="Text", uConc="mg/L")

#>           b0         CMAX        CMAXD         TMAX         TLAG         CLST 
#>    2.3687851   10.5000000    0.0328125    1.1200000    0.0000000    3.2800000 
#>        CLSTP         TLST       LAMZHL         LAMZ       LAMZLL       LAMZUL 
#>    3.2801465   24.3700000   14.3043776    0.0484570    9.0500000   24.3700000 
#>      LAMZNPT       CORRXY           R2        R2ADJ           C0       AUCLST 
#>    3.0000000   -0.9999999    0.9999997    0.9999995           NA  148.9230500 
#>       AUCALL       AUCIFO      AUCIFOD       AUCIFP      AUCIFPD       AUCPEO 
#>  148.9230500  216.6119330    0.6769123  216.6149558    0.6769217   31.2489169 
#>       AUCPEP      AUCPBEO      AUCPBEP      AUMCLST      AUMCIFO      AUMCIFP 
#>   31.2498763           NA           NA 1459.0711035 4505.5348194 4505.6708646 
#>      AUMCPEO      AUMCPEP     MRTIVLST     MRTIVIFO     MRTIVIFP     MRTEVLST 
#>   67.6160287   67.6170065           NA           NA           NA    9.7974834 
#>     MRTEVIFO     MRTEVIFP          VZO          VZP         VZFO         VZFP 
#>   20.8000305   20.8003683           NA           NA   30.4867482   30.4863228 
#>          CLO          CLP         CLFO         CLFP         VSSO         VSSP 
#>           NA           NA    1.4772963    1.4772757           NA           NA 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Bolus", report="Text", uConc="mg/L")

#>          b0        CMAX       CMAXD        TMAX        TLAG        CLST 
#> -1.72421060  1.50000000  0.06000000  0.25000000          NA  0.05000000 
#>       CLSTP        TLST      LAMZHL        LAMZ      LAMZLL      LAMZUL 
#>  0.05024851  8.00000000  4.37812701  0.15832048  5.00000000  8.00000000 
#>     LAMZNPT      CORRXY          R2       R2ADJ          C0      AUCLST 
#>  3.00000000 -0.99853229  0.99706673  0.99413345  2.39361702  2.04045213 
#>      AUCALL      AUCIFO     AUCIFOD      AUCIFP     AUCIFPD      AUCPEO 
#>  2.04045213  2.35626723  0.09425069  2.35783688  0.09431348 13.40319561 
#>      AUCPEP     AUCPBEO     AUCPBEP     AUMCLST     AUMCIFO     AUMCIFP 
#> 13.46084419 20.65564214 20.64189142  3.27125000  7.79255448  7.81502594 
#>     AUMCPEO     AUMCPEP    MRTIVLST    MRTIVIFO    MRTIVIFP    MRTEVLST 
#> 58.02082606 58.14153371  1.60319860  3.30716074  3.31448966          NA 
#>    MRTEVIFO    MRTEVIFP         VZO         VZP        VZFO        VZFP 
#>          NA          NA 67.01597804 66.97136466          NA          NA 
#>         CLO         CLP        CLFO        CLFP        VSSO        VSSP 
#> 10.61000197 10.60293876          NA          NA 35.08898193 35.14333088 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Infusion", dur=0.25, report="Text", uConc="mg/L")

#>          b0        CMAX       CMAXD        TMAX        TLAG        CLST 
#> -1.72421060  1.50000000  0.06000000  0.25000000  0.00000000  0.05000000 
#>       CLSTP        TLST      LAMZHL        LAMZ      LAMZLL      LAMZUL 
#>  0.05024851  8.00000000  4.37812701  0.15832048  5.00000000  8.00000000 
#>     LAMZNPT      CORRXY          R2       R2ADJ          C0      AUCLST 
#>  3.00000000 -0.99853229  0.99706673  0.99413345          NA  1.74125000 
#>      AUCALL      AUCIFO     AUCIFOD      AUCIFP     AUCIFPD      AUCPEO 
#>  1.74125000  2.05706511  0.08228260  2.05863475  0.08234539 15.35270349 
#>      AUCPEP     AUCPBEO     AUCPBEP     AUMCLST     AUMCIFO     AUMCIFP 
#> 15.41724428          NA          NA  3.27125000  7.79255448  7.81502594 
#>     AUMCPEO     AUMCPEP    MRTIVLST    MRTIVIFO    MRTIVIFP    MRTEVLST 
#> 58.02082606 58.14153371  1.75367911  3.66319049  3.67121783          NA 
#>    MRTEVIFO    MRTEVIFP         VZO         VZP        VZFO        VZFP 
#>          NA          NA 76.76351746 76.70498779          NA          NA 
#>         CLO         CLP        CLFO        CLFP        VSSO        VSSP 
#> 12.15323712 12.14397067          NA          NA 44.51962266 44.58316170 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        

iAUC = data.frame(Name=c("AUC[0-12h]","AUC[0-24h]"), Start=c(0,0), End=c(12,24)) ; iAUC

#>         Name Start End
#> 1 AUC[0-12h]     0  12
#> 2 AUC[0-24h]     0  24

IndiNCA(Theoph[Theoph$Subject==1,"Time"], Theoph[Theoph$Subject==1, "conc"], dose=320,
        iAUC=iAUC, uConc="mg/L")

#>           b0         CMAX        CMAXD         TMAX         TLAG         CLST 
#>    2.3687851   10.5000000    0.0328125    1.1200000    0.0000000    3.2800000 
#>        CLSTP         TLST       LAMZHL         LAMZ       LAMZLL       LAMZUL 
#>    3.2801465   24.3700000   14.3043776    0.0484570    9.0500000   24.3700000 
#>      LAMZNPT       CORRXY           R2        R2ADJ           C0       AUCLST 
#>    3.0000000   -0.9999999    0.9999997    0.9999995           NA  148.9230500 
#>       AUCALL       AUCIFO      AUCIFOD       AUCIFP      AUCIFPD       AUCPEO 
#>  148.9230500  216.6119330    0.6769123  216.6149558    0.6769217   31.2489169 
#>       AUCPEP      AUCPBEO      AUCPBEP      AUMCLST      AUMCIFO      AUMCIFP 
#>   31.2498763           NA           NA 1459.0711035 4505.5348194 4505.6708646 
#>      AUMCPEO      AUMCPEP     MRTIVLST     MRTIVIFO     MRTIVIFP     MRTEVLST 
#>   67.6160287   67.6170065           NA           NA           NA    9.7974834 
#>     MRTEVIFO     MRTEVIFP          VZO          VZP         VZFO         VZFP 
#>   20.8000305   20.8003683           NA           NA   30.4867482   30.4863228 
#>          CLO          CLP         CLFO         CLFP         VSSO         VSSP 
#>           NA           NA    1.4772963    1.4772757           NA           NA 
#>   AUC[0-12h]   AUC[0-24h] 
#>   91.7355220  147.6945866 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        
#> [49] "h*mg/L"    "h*mg/L"   

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Bolus", iAUC=iAUC, uConc="mg/L")

#>          b0        CMAX       CMAXD        TMAX        TLAG        CLST 
#> -1.72421060  1.50000000  0.06000000  0.25000000          NA  0.05000000 
#>       CLSTP        TLST      LAMZHL        LAMZ      LAMZLL      LAMZUL 
#>  0.05024851  8.00000000  4.37812701  0.15832048  5.00000000  8.00000000 
#>     LAMZNPT      CORRXY          R2       R2ADJ          C0      AUCLST 
#>  3.00000000 -0.99853229  0.99706673  0.99413345  2.39361702  2.04045213 
#>      AUCALL      AUCIFO     AUCIFOD      AUCIFP     AUCIFPD      AUCPEO 
#>  2.04045213  2.35626723  0.09425069  2.35783688  0.09431348 13.40319561 
#>      AUCPEP     AUCPBEO     AUCPBEP     AUMCLST     AUMCIFO     AUMCIFP 
#> 13.46084419 20.65564214 20.64189142  3.27125000  7.79255448  7.81502594 
#>     AUMCPEO     AUMCPEP    MRTIVLST    MRTIVIFO    MRTIVIFP    MRTEVLST 
#> 58.02082606 58.14153371  1.60319860  3.30716074  3.31448966          NA 
#>    MRTEVIFO    MRTEVIFP         VZO         VZP        VZFO        VZFP 
#>          NA          NA 67.01597804 66.97136466          NA          NA 
#>         CLO         CLP        CLFO        CLFP        VSSO        VSSP 
#> 10.61000197 10.60293876          NA          NA 35.08898193 35.14333088 
#>  AUC[0-12h]  AUC[0-24h] 
#>  2.18894711  2.33163353 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        
#> [49] "h*mg/L"    "h*mg/L"   

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Infusion", dur=0.25, iAUC=iAUC, uConc="mg/L")

#>          b0        CMAX       CMAXD        TMAX        TLAG        CLST 
#> -1.72421060  1.50000000  0.06000000  0.25000000  0.00000000  0.05000000 
#>       CLSTP        TLST      LAMZHL        LAMZ      LAMZLL      LAMZUL 
#>  0.05024851  8.00000000  4.37812701  0.15832048  5.00000000  8.00000000 
#>     LAMZNPT      CORRXY          R2       R2ADJ          C0      AUCLST 
#>  3.00000000 -0.99853229  0.99706673  0.99413345          NA  1.74125000 
#>      AUCALL      AUCIFO     AUCIFOD      AUCIFP     AUCIFPD      AUCPEO 
#>  1.74125000  2.05706511  0.08228260  2.05863475  0.08234539 15.35270349 
#>      AUCPEP     AUCPBEO     AUCPBEP     AUMCLST     AUMCIFO     AUMCIFP 
#> 15.41724428          NA          NA  3.27125000  7.79255448  7.81502594 
#>     AUMCPEO     AUMCPEP    MRTIVLST    MRTIVIFO    MRTIVIFP    MRTEVLST 
#> 58.02082606 58.14153371  1.75367911  3.66319049  3.67121783          NA 
#>    MRTEVIFO    MRTEVIFP         VZO         VZP        VZFO        VZFP 
#>          NA          NA 76.76351746 76.70498779          NA          NA 
#>         CLO         CLP        CLFO        CLFP        VSSO        VSSP 
#> 12.15323712 12.14397067          NA          NA 44.51962266 44.58316170 
#>  AUC[0-12h]  AUC[0-24h] 
#>  1.88974498  2.03243140 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        
#> [49] "h*mg/L"    "h*mg/L"