Noncompartmental analysis for a dataset with multiple subjects

conduct noncompartmental analysis for many subjects in a data table

NCA(concData, id, Time, conc, trt="", fit = "Linear", dose = 0,
           adm = "Extravascular", dur = 0, report = "Table", iAUC = "",
           uTime = "h", uConc = "ug/L", uDose = "mg")

Arguments

concData
name of data table containing time-concentration data of multiple subjects

id
column name for subject ID

Time
column name for the time

conc
column name for the concentration

trt
column name for the treatment code. This is useful for crossover study like bioequivalence trial.

fit
one of "Linear" or "Log" to indicate the way to calculate AUC

dose
administered dose. One should be careful for the unit. This can be a vector containing dose for each subject in order.

adm
one of "Bolus" or "Infusion" or "Extravascular" to indicate drug administration mode

dur
infusion duration for constant infusion, otherwise 0. This can be a vector containing values for each subject in order.

report
either of "Table" or "Text" to specify the type of return value

iAUC
data.frame with three columns, "Name", "Start", "End" to specify partial interval AUC

uTime
unit of time

uConc
unit of concentration

uDose
unit of dose

Details

This function calls IndiNCA repeatedly to do NCA for each subject. If you specify Report="Text", this function returns in free text format to be used in a report file.

Value

CMAX

maximum concentration, Cmax

CMAXD

dose normalized Cmax, CMAX / Dose, Cmax / Dose

TMAX

time of maximum concentration, Tmax

TLAG

time to observe the first non-zero concentration, for extravascular administration only

CLST

last positive concentration observed, Clast

CLSTP

last positive concentration predicted, Clast_pred

TLST

time of last positive concentration, Tlast

LAMZHL

half-life by lambda z, ln(2)/LAMZ

LAMZ

lambda_z negative of best fit terminal slope

LAMZLL

earliest time for LAMZ

LAMZUL

last time for LAMZ

LAMZNPT

number of points for LAMZ

CORRXY

correlation of log(concentration) and time

R2

R-squared

R2ADJ

R-squared adjusted

C0

back extrapolated concentration at time 0, for bolus intravascular administration only

AUCLST

AUC from 0 to TLST

AUCALL

AUC using all the given points, including trailing zero concentrations

AUCIFO

AUC infinity observed

AUCIFOD

AUCIFO / Dose

AUCIFP

AUC infinity predicted using CLSTP instead of CLST

AUCIFPD

AUCIFP / Dose

AUCPEO

AUC % extrapolation observed

AUCPEP

AUC % extrapolated for AUCIFP

AUCPBEO

AUC % back extrapolation observed, for bolus IV administration only

AUCPBEP

AUC % back extrapolation predicted with AUCIFP, for bolus IV administration only

AUMCLST

AUMC to the TLST

AUMCIFO

AUMC infinity observed using CLST

AUMCIFP

AUMC infinity determined by CLSTP

AUMCPEO

AUMC % extrapolated observed

AUMCPEP

AUMC % extrapolated predicted

MRTIVLST

mean residence time (MRT) to TLST, for intravascular administration

MRTIVIFO

mean residence time (MRT) infinity using CLST, for intravascular administration

MRTIVIFP

mean residence time (MRT) infinity using CLSTP, for intravascular administration

MRTEVLST

mean residence time (MRT) to TLST, for extravascular administration

MRTEVIFO

mean residence time (MRT) infinity using CLST, for extravascular administration

MRTEVIFP

mean residence time (MRT) infinity using CLSTP, for extravascular administration

VZO

volume of distribution determined by LAMZ and AUCIFO, for intravascular administration

VZP

volume of distribution determined by LAMZ and AUCIFP, for intravascular administration

VZFO

VZO for extravascular administration, VZO/F, F is bioavailability

VZFP

VZP for extravascular administration, VZP/F, F is bioavailability

CLO

clearance using AUCIFO, for intravascular administration

CLP

clearance using AUCIFP, for intravascular administration

CLFO

CLO for extravascular administration, CLO/F, F is bioavailability

CLFP

CLP for extravascular administration, CLP/F, F is bioavailability

VSSO

volume of distribution at steady state using CLST, for intravascular administration only

VSSP

volume of distribution at stead state using CLSTP, for intravascular administration only

References

Gabrielsson J, Weiner D. Pharmacokinetic and Pharmacodynamic Data Analysis - Concepts and Applications. 5th ed. 2016.

Shargel L, Yu A. Applied Biopharmaceutics and Pharmacokinetics. 7th ed. 2015.

Rowland M, Tozer TN. Clinical Pharmacokinetics and Pharmacodynamics - Concepts and Applications. 4th ed. 2011.

Gibaldi M, Perrier D. Pharmacokinetics. 2nd ed. revised and expanded. 1982.

See also

Examples

# Theoph and Indometh data: dose in mg, conc in mg/L, time in h
NCA(Theoph, "Subject", "Time", "conc", dose=320, uConc="mg/L")

#>    Subject  CMAX      CMAXD TMAX TLAG CLST     CLSTP  TLST    LAMZHL       LAMZ
#> 1        1 10.50 0.03281250 1.12    0 3.28 3.2801465 24.37 14.304378 0.04845700
#> 2        2  8.33 0.02603125 1.92    0 0.90 0.8886398 24.30  6.659342 0.10408644
#> 3        3  8.20 0.02562500 1.02    0 1.05 1.0550967 24.17  6.766087 0.10244431
#> 4        4  8.60 0.02687500 1.07    0 1.15 1.1564216 24.65  6.981247 0.09928702
#> 5        5 11.40 0.03562500 1.00    0 1.57 1.5556951 24.35  8.002264 0.08661888
#> 6        6  6.44 0.02012500 1.15    0 0.92 0.9412712 23.85  7.894998 0.08779574
#> 7        7  7.09 0.02215625 3.48    0 1.15 1.1607192 24.22  7.846668 0.08833650
#> 8        8  7.56 0.02362500 2.02    0 1.25 1.2285268 24.12  8.510038 0.08145054
#> 9        9  9.03 0.02821875 0.63    0 1.12 1.1164831 24.43  8.405999 0.08245863
#> 10      10 10.21 0.03190625 3.55    0 2.42 2.4136923 23.70  9.246916 0.07495982
#> 11      11  8.00 0.02500000 0.98    0 0.86 0.8598066 24.08  7.261237 0.09545856
#> 12      12  9.75 0.03046875 3.52    0 1.17 1.1755390 24.15  6.286508 0.11025949
#>    LAMZLL LAMZUL LAMZNPT     CORRXY        R2     R2ADJ    AUCLST    AUCALL
#> 1    9.05  24.37       3 -0.9999999 0.9999997 0.9999995 148.92305 148.92305
#> 2    7.03  24.30       4 -0.9985967 0.9971954 0.9957931  91.52680  91.52680
#> 3    9.00  24.17       3 -0.9996624 0.9993250 0.9986499  99.28650  99.28650
#> 4    9.02  24.65       3 -0.9994619 0.9989241 0.9978483 106.79630 106.79630
#> 5    7.02  24.35       4 -0.9993234 0.9986472 0.9979708 121.29440 121.29440
#> 6    2.03  23.85       7 -0.9991203 0.9982413 0.9978896  73.77555  73.77555
#> 7    6.98  24.22       4 -0.9993349 0.9986702 0.9980053  90.75340  90.75340
#> 8    3.53  24.12       6 -0.9954961 0.9910124 0.9887655  88.55995  88.55995
#> 9    8.80  24.43       3 -0.9997218 0.9994437 0.9988873  86.32615  86.32615
#> 10   9.38  23.70       3 -0.9997543 0.9995087 0.9990174 138.36810 138.36810
#> 11   9.03  24.08       3 -0.9999991 0.9999983 0.9999965  80.09360  80.09360
#> 12   9.03  24.15       3 -0.9996984 0.9993968 0.9987936 119.97750 119.97750
#>       AUCIFO   AUCIFOD    AUCPEO    AUCIFP   AUCIFPD    AUCPEP   AUMCLST
#> 1  216.61193 0.6769123 216.61496 0.6769217 31.248917 31.249876 1459.0711
#> 2  100.17346 0.3130421 100.06432 0.3127010  8.631687  8.532030  706.5866
#> 3  109.53597 0.3422999 109.58572 0.3424554  9.357173  9.398325  803.1859
#> 4  118.37888 0.3699340 118.44356 0.3701361  9.784331  9.833594  901.0842
#> 5  139.41978 0.4356868 139.25463 0.4351707 13.000579 12.897403 1017.1143
#> 6   84.25442 0.2632951  84.49670 0.2640522 12.437174 12.688246  609.1524
#> 7  103.77180 0.3242869 103.89315 0.3246661 12.545221 12.647366  782.4199
#> 8  103.90669 0.3247084 103.64305 0.3238845 14.769730 14.552931  739.5346
#> 9   99.90872 0.3122147  99.86607 0.3120815 13.594978 13.558076  705.2296
#> 10 170.65206 0.5332877 170.56791 0.5330247 18.918002 18.878001 1278.1800
#> 11  89.10274 0.2784461  89.10072 0.2784397 10.110962 10.108918  617.2422
#> 12 130.58883 0.4080901 130.63907 0.4082471  8.125757  8.161087  977.8807
#>      AUMCIFO   AUMCPEO  AUMCIFP  AUMCPEP MRTEVLST  MRTEVIFO  MRTEVIFP     VZFO
#> 1  4505.5348 4505.6709 67.61603 67.61701 9.797483 20.800031 20.800368 30.48675
#> 2   999.7723  996.0716 29.32525 29.06267 7.719996  9.980411  9.954313 30.69044
#> 3  1150.9648 1152.6529 30.21629 30.31850 8.089578 10.507642 10.518276 28.51710
#> 4  1303.2524 1305.4981 30.85881 30.97775 8.437410 11.009163 11.022112 27.22596
#> 5  1667.7216 1661.7937 39.01174 38.79419 8.385501 11.961873 11.933490 26.49799
#> 6   978.4285  986.9665 37.74176 38.28034 8.256833 11.612785 11.680533 43.25973
#> 7  1245.0984 1249.4111 37.16000 37.37691 8.621383 11.998427 12.025924 34.90844
#> 8  1298.1158 1288.5201 43.03015 42.60589 8.350666 12.493092 12.432287 37.81051
#> 9  1201.7715 1200.2124 41.31750 41.24126 8.169363 12.028695 12.018220 38.84279
#> 10 2473.9934 2470.8765 48.33535 48.27018 9.237534 14.497296 14.486174 25.01554
#> 11  928.5600  928.4900 33.52694 33.52193 7.706511 10.421227 10.420679 37.62219
#> 12 1330.3840 1332.0528 26.49636 26.58844 8.150534 10.187579 10.196436 22.22429
#>        VZFP     CLFO     CLFP
#> 1  30.48632 1.477296 1.477276
#> 2  30.72392 3.194459 3.197943
#> 3  28.50415 2.921415 2.920088
#> 4  27.21110 2.703185 2.701709
#> 5  26.52942 2.295227 2.297949
#> 6  43.13569 3.798020 3.787130
#> 7  34.86767 3.083689 3.080088
#> 8  37.90669 3.079686 3.087520
#> 9  38.85938 3.202924 3.204292
#> 10 25.02788 1.875160 1.876086
#> 11 37.62304 3.591360 3.591441
#> 12 22.21575 2.450439 2.449497

NCA(Indometh, "Subject", "time", "conc", dose=25, adm="Bolus", uConc="mg/L")

#>   Subject CMAX  CMAXD TMAX   C0    AUCPBEO AUCPBEP     CLST     CLSTP TLST
#> 1       1 1.50 0.0600 0.25 0.05 0.05024851       8 4.378127 0.1583205 5.00
#> 2       2 2.03 0.0812 0.25 0.08 0.07475591       8 2.293063 0.3022800 0.75
#> 3       3 2.72 0.1088 0.25 0.08 0.03997258       8 1.642947 0.4218926 0.50
#> 4       4 1.85 0.0740 0.25 0.07 0.03196011       8 1.521910 0.4554455 0.25
#> 5       5 2.05 0.0820 0.25 0.06 0.04957149       8 2.742446 0.2527478 1.00
#> 6       6 2.31 0.0924 0.25 0.09 0.05624247       8 1.960699 0.3535205 0.75
#>   LAMZHL LAMZ     LAMZLL    LAMZUL   LAMZNPT   CORRXY       R2    R2ADJ
#> 1      8    3 -0.9985323 0.9970667 0.9941335 2.393617 2.040452 2.040452
#> 2      8    9 -0.9734830 0.9476691 0.9401933 2.528160 3.248520 3.248520
#> 3      8   10 -0.9358558 0.8758261 0.8603043 4.965369 3.554421 3.554421
#> 4      8   11 -0.9342510 0.8728249 0.8586943 2.462230 2.785279 2.785279
#> 5      8    8 -0.9355449 0.8752442 0.8544516 4.040865 2.458858 2.458858
#> 6      8    9 -0.9507649 0.9039538 0.8902329 3.705625 3.335703 3.335703
#>     AUCLST     AUCALL   AUCIFO    AUCIFOD    AUCPEO    AUCIFP  AUCIFPD   AUCPEP
#> 1 2.356267 0.09425069 2.357837 0.09431348 13.403196 13.460844 20.65564 20.64189
#> 2 3.513175 0.14052701 3.495827 0.13983307  7.533221  7.074344 16.21809 16.29857
#> 3 3.744043 0.14976171 3.649167 0.14596668  5.064624  2.596369 25.65866 26.32577
#> 4 2.938974 0.11755898 2.855452 0.11421808  5.229568  2.457520 18.34071 18.87718
#> 5 2.696249 0.10784996 2.654988 0.10619954  8.804484  7.387236 28.23768 28.67652
#> 6 3.590285 0.14361141 3.494796 0.13979183  7.090860  4.552269 20.94411 21.51637
#>    AUMCLST  AUMCIFO  AUMCPEO  AUMCIFP  AUMCPEP MRTIVLST MRTIVIFO MRTIVIFP
#> 1 3.271250 7.792554 7.815026 58.02083 58.14153 1.603199 3.307161 3.314490
#> 2 6.398750 9.391522 9.195343 31.86674 30.41314 1.969743 2.673229 2.630377
#> 3 5.006250 6.972678 5.988790 28.20191 16.40632 1.408457 1.862339 1.641139
#> 4 4.381875 5.948903 5.097338 26.34146 14.03601 1.573227 2.024142 1.785125
#> 5 3.707500 6.545866 6.052534 43.36120 38.74467 1.507814 2.427768 2.279684
#> 6 5.532500 8.289291 7.255264 33.25726 23.74502 1.658571 2.308811 2.076019
#>        VZO      VZP       CLO       CLP     VSSO     VSSP
#> 1 67.01598 66.97136 10.610002 10.602939 35.08898 35.14333
#> 2 23.54132 23.65814  7.116070  7.151384 19.02289 18.81084
#> 3 15.82695 16.23844  6.677274  6.850879 12.43535 11.24324
#> 4 18.67703 19.22334  8.506369  8.755181 17.21810 15.62909
#> 5 36.68535 37.25547  9.272141  9.416237 22.51060 21.46604
#> 6 19.69683 20.23502  6.963235  7.153494 16.07680 14.85079

iAUC = data.frame(Name=c("AUC[0-12h]","AUC[0-24h]"), Start=c(0,0), End=c(12,24)) ; iAUC

#>         Name Start End
#> 1 AUC[0-12h]     0  12
#> 2 AUC[0-24h]     0  24

NCA(Theoph, "Subject", "Time", "conc", dose=320, iAUC=iAUC, uConc="mg/L")

#>    Subject  CMAX      CMAXD TMAX TLAG CLST     CLSTP  TLST    LAMZHL       LAMZ
#> 1        1 10.50 0.03281250 1.12    0 3.28 3.2801465 24.37 14.304378 0.04845700
#> 2        2  8.33 0.02603125 1.92    0 0.90 0.8886398 24.30  6.659342 0.10408644
#> 3        3  8.20 0.02562500 1.02    0 1.05 1.0550967 24.17  6.766087 0.10244431
#> 4        4  8.60 0.02687500 1.07    0 1.15 1.1564216 24.65  6.981247 0.09928702
#> 5        5 11.40 0.03562500 1.00    0 1.57 1.5556951 24.35  8.002264 0.08661888
#> 6        6  6.44 0.02012500 1.15    0 0.92 0.9412712 23.85  7.894998 0.08779574
#> 7        7  7.09 0.02215625 3.48    0 1.15 1.1607192 24.22  7.846668 0.08833650
#> 8        8  7.56 0.02362500 2.02    0 1.25 1.2285268 24.12  8.510038 0.08145054
#> 9        9  9.03 0.02821875 0.63    0 1.12 1.1164831 24.43  8.405999 0.08245863
#> 10      10 10.21 0.03190625 3.55    0 2.42 2.4136923 23.70  9.246916 0.07495982
#> 11      11  8.00 0.02500000 0.98    0 0.86 0.8598066 24.08  7.261237 0.09545856
#> 12      12  9.75 0.03046875 3.52    0 1.17 1.1755390 24.15  6.286508 0.11025949
#>    LAMZLL LAMZUL LAMZNPT     CORRXY        R2     R2ADJ    AUCLST    AUCALL
#> 1    9.05  24.37       3 -0.9999999 0.9999997 0.9999995 148.92305 148.92305
#> 2    7.03  24.30       4 -0.9985967 0.9971954 0.9957931  91.52680  91.52680
#> 3    9.00  24.17       3 -0.9996624 0.9993250 0.9986499  99.28650  99.28650
#> 4    9.02  24.65       3 -0.9994619 0.9989241 0.9978483 106.79630 106.79630
#> 5    7.02  24.35       4 -0.9993234 0.9986472 0.9979708 121.29440 121.29440
#> 6    2.03  23.85       7 -0.9991203 0.9982413 0.9978896  73.77555  73.77555
#> 7    6.98  24.22       4 -0.9993349 0.9986702 0.9980053  90.75340  90.75340
#> 8    3.53  24.12       6 -0.9954961 0.9910124 0.9887655  88.55995  88.55995
#> 9    8.80  24.43       3 -0.9997218 0.9994437 0.9988873  86.32615  86.32615
#> 10   9.38  23.70       3 -0.9997543 0.9995087 0.9990174 138.36810 138.36810
#> 11   9.03  24.08       3 -0.9999991 0.9999983 0.9999965  80.09360  80.09360
#> 12   9.03  24.15       3 -0.9996984 0.9993968 0.9987936 119.97750 119.97750
#>       AUCIFO   AUCIFOD    AUCPEO    AUCIFP   AUCIFPD    AUCPEP   AUMCLST
#> 1  216.61193 0.6769123 216.61496 0.6769217 31.248917 31.249876 1459.0711
#> 2  100.17346 0.3130421 100.06432 0.3127010  8.631687  8.532030  706.5866
#> 3  109.53597 0.3422999 109.58572 0.3424554  9.357173  9.398325  803.1859
#> 4  118.37888 0.3699340 118.44356 0.3701361  9.784331  9.833594  901.0842
#> 5  139.41978 0.4356868 139.25463 0.4351707 13.000579 12.897403 1017.1143
#> 6   84.25442 0.2632951  84.49670 0.2640522 12.437174 12.688246  609.1524
#> 7  103.77180 0.3242869 103.89315 0.3246661 12.545221 12.647366  782.4199
#> 8  103.90669 0.3247084 103.64305 0.3238845 14.769730 14.552931  739.5346
#> 9   99.90872 0.3122147  99.86607 0.3120815 13.594978 13.558076  705.2296
#> 10 170.65206 0.5332877 170.56791 0.5330247 18.918002 18.878001 1278.1800
#> 11  89.10274 0.2784461  89.10072 0.2784397 10.110962 10.108918  617.2422
#> 12 130.58883 0.4080901 130.63907 0.4082471  8.125757  8.161087  977.8807
#>      AUMCIFO   AUMCPEO  AUMCIFP  AUMCPEP MRTEVLST  MRTEVIFO  MRTEVIFP     VZFO
#> 1  4505.5348 4505.6709 67.61603 67.61701 9.797483 20.800031 20.800368 30.48675
#> 2   999.7723  996.0716 29.32525 29.06267 7.719996  9.980411  9.954313 30.69044
#> 3  1150.9648 1152.6529 30.21629 30.31850 8.089578 10.507642 10.518276 28.51710
#> 4  1303.2524 1305.4981 30.85881 30.97775 8.437410 11.009163 11.022112 27.22596
#> 5  1667.7216 1661.7937 39.01174 38.79419 8.385501 11.961873 11.933490 26.49799
#> 6   978.4285  986.9665 37.74176 38.28034 8.256833 11.612785 11.680533 43.25973
#> 7  1245.0984 1249.4111 37.16000 37.37691 8.621383 11.998427 12.025924 34.90844
#> 8  1298.1158 1288.5201 43.03015 42.60589 8.350666 12.493092 12.432287 37.81051
#> 9  1201.7715 1200.2124 41.31750 41.24126 8.169363 12.028695 12.018220 38.84279
#> 10 2473.9934 2470.8765 48.33535 48.27018 9.237534 14.497296 14.486174 25.01554
#> 11  928.5600  928.4900 33.52694 33.52193 7.706511 10.421227 10.420679 37.62219
#> 12 1330.3840 1332.0528 26.49636 26.58844 8.150534 10.187579 10.196436 22.22429
#>        VZFP     CLFO     CLFP AUC[0-12h] AUC[0-24h]
#> 1  30.48632 1.477296 1.477276   91.73552  147.69459
#> 2  30.72392 3.194459 3.197943   67.48030   91.24908
#> 3  28.50415 2.921415 2.920088   70.17971   99.10481
#> 4  27.21110 2.703185 2.701709   73.05115  105.99811
#> 5  26.52942 2.295227 2.297949   84.61490  120.73101
#> 6  43.13569 3.798020 3.787130   51.75887   73.91422
#> 7  34.86767 3.083689 3.080088   62.09875   90.49567
#> 8  37.90669 3.079686 3.087520   62.71486   88.40890
#> 9  38.85938 3.202924 3.204292   60.12123   85.82985
#> 10 25.02788 1.875160 1.876086   90.81742  139.08507
#> 11 37.62304 3.591360 3.591441   58.53963   80.02431
#> 12 22.21575 2.450439 2.449497   85.02136  119.79884

NCA(Indometh, "Subject", "time", "conc", dose=25, adm="Bolus", iAUC=iAUC, uConc="mg/L")

#>   Subject CMAX  CMAXD TMAX   C0    AUCPBEO AUCPBEP     CLST     CLSTP TLST
#> 1       1 1.50 0.0600 0.25 0.05 0.05024851       8 4.378127 0.1583205 5.00
#> 2       2 2.03 0.0812 0.25 0.08 0.07475591       8 2.293063 0.3022800 0.75
#> 3       3 2.72 0.1088 0.25 0.08 0.03997258       8 1.642947 0.4218926 0.50
#> 4       4 1.85 0.0740 0.25 0.07 0.03196011       8 1.521910 0.4554455 0.25
#> 5       5 2.05 0.0820 0.25 0.06 0.04957149       8 2.742446 0.2527478 1.00
#> 6       6 2.31 0.0924 0.25 0.09 0.05624247       8 1.960699 0.3535205 0.75
#>   LAMZHL LAMZ     LAMZLL    LAMZUL   LAMZNPT   CORRXY       R2    R2ADJ
#> 1      8    3 -0.9985323 0.9970667 0.9941335 2.393617 2.040452 2.040452
#> 2      8    9 -0.9734830 0.9476691 0.9401933 2.528160 3.248520 3.248520
#> 3      8   10 -0.9358558 0.8758261 0.8603043 4.965369 3.554421 3.554421
#> 4      8   11 -0.9342510 0.8728249 0.8586943 2.462230 2.785279 2.785279
#> 5      8    8 -0.9355449 0.8752442 0.8544516 4.040865 2.458858 2.458858
#> 6      8    9 -0.9507649 0.9039538 0.8902329 3.705625 3.335703 3.335703
#>     AUCLST     AUCALL   AUCIFO    AUCIFOD    AUCPEO    AUCIFP  AUCIFPD   AUCPEP
#> 1 2.356267 0.09425069 2.357837 0.09431348 13.403196 13.460844 20.65564 20.64189
#> 2 3.513175 0.14052701 3.495827 0.13983307  7.533221  7.074344 16.21809 16.29857
#> 3 3.744043 0.14976171 3.649167 0.14596668  5.064624  2.596369 25.65866 26.32577
#> 4 2.938974 0.11755898 2.855452 0.11421808  5.229568  2.457520 18.34071 18.87718
#> 5 2.696249 0.10784996 2.654988 0.10619954  8.804484  7.387236 28.23768 28.67652
#> 6 3.590285 0.14361141 3.494796 0.13979183  7.090860  4.552269 20.94411 21.51637
#>    AUMCLST  AUMCIFO  AUMCPEO  AUMCIFP  AUMCPEP MRTIVLST MRTIVIFO MRTIVIFP
#> 1 3.271250 7.792554 7.815026 58.02083 58.14153 1.603199 3.307161 3.314490
#> 2 6.398750 9.391522 9.195343 31.86674 30.41314 1.969743 2.673229 2.630377
#> 3 5.006250 6.972678 5.988790 28.20191 16.40632 1.408457 1.862339 1.641139
#> 4 4.381875 5.948903 5.097338 26.34146 14.03601 1.573227 2.024142 1.785125
#> 5 3.707500 6.545866 6.052534 43.36120 38.74467 1.507814 2.427768 2.279684
#> 6 5.532500 8.289291 7.255264 33.25726 23.74502 1.658571 2.308811 2.076019
#>        VZO      VZP       CLO       CLP     VSSO     VSSP AUC[0-12h] AUC[0-24h]
#> 1 67.01598 66.97136 10.610002 10.602939 35.08898 35.14333   2.188947   2.331634
#> 2 23.54132 23.65814  7.116070  7.151384 19.02289 18.81084   3.429231   3.507581
#> 3 15.82695 16.23844  6.677274  6.850879 12.43535 11.24324   3.676377   3.726268
#> 4 18.67703 19.22334  8.506369  8.755181 17.21810 15.62909   2.884797   2.924002
#> 5 36.68535 37.25547  9.272141  9.416237 22.51060 21.46604   2.598511   2.682264
#> 6 19.69683 20.23502  6.963235  7.153494 16.07680 14.85079   3.497732   3.570236

writeLines(NCA(Theoph, "Subject", "Time", "conc", dose=320, report="Text", uConc="mg/L"),
           "Theoph_Linear_CoreOutput.txt")
writeLines(NCA(Theoph, "Subject", "Time", "conc", dose=320, fit="Log", report="Text",
           uConc="mg/L"), "Theoph_Log_CoreOutput.txt")
writeLines(NCA(Indometh, "Subject", "time", "conc", dose=25, adm="Bolus", report="Text",
           uConc="mg/L"), "Indometh_Bolus_Linear_CoreOutput.txt")
writeLines(NCA(Indometh, "Subject", "time", "conc", dose=25, adm="Bolus", fit="Log",
           report="Text", uConc="mg/L"), "Indometh_Bolus_Log_CoreOutput.txt")
writeLines(NCA(Indometh, "Subject", "time", "conc", dose=25, adm="Infusion", dur=0.25,
           report="Text", uConc="mg/L"), "Indometh_Infusion_Linear_CoreOutput.txt")
writeLines(NCA(Indometh, "Subject", "time", "conc", dose=25, adm="Infusion", dur=0.25,
           fit="Log", report="Text", uConc="mg/L"), "Indometh_Infusion_Log_CoreOutput.txt")