Get the Slope of regression log(y) ~ x

It calculates the slope with linear regression of log(y) ~ x

Slope(x, y)

Arguments

x
vector values of independent variable, usually time

y
vector values of dependent variable, usually concentration

Details

With time-concentration curve, you frequently need to estimate slope in log(concentration) ~ time. This function is usually called by BestSlope function and you seldom need to call this function directly.

Value

R2

R-squared

R2ADJ

adjusted R-squared

LAMZNPT

number of points used for slope

LAMZ

negative of slope, lambda_z

b0

intercept of regression line

CORRXY

correlation of log(y) and x

LAMZLL

earliest x for lambda_z

LAMZUL

last x for lambda_z

CLSTP

predicted y value at last point, predicted concentration for the last time point

See also

BestSlope

Examples

Slope(Indometh[Indometh$Subject==1, "time"],Indometh[Indometh$Subject==1, "conc"])

#>         R2      R2ADJ    LAMZNPT       LAMZ         b0     CORRXY     LAMZLL 
#>  0.5340746  0.4823051 11.0000000  0.1332090  0.8108532 -0.7308041  0.2500000 
#>     LAMZUL      CLSTP 
#>  8.0000000  0.7750570