Simplest NCA

This is the work-horse function for NCA.

sNCA(x, y, dose = 0, adm = "Extravascular", dur = 0, doseUnit = "mg", timeUnit = "h",
     concUnit = "ug/L", iAUC = "", down = "Linear", MW = 0, returnNA = TRUE)

Arguments

x
usually time

y
usually concentration

dose
given amount

adm
one of "Bolus" or "Infusion" or "Extravascular" to indicate drug administration mode

dur
duration of infusion

doseUnit
unit of dose

timeUnit
unit of time

concUnit
unit of concentration

iAUC
interval AUCs to calculate

down
either of "Linear" or "Log" to indicate the way to calculate AUC and AUMC

MW
molecular weight of the drug

returnNA
if returnNA is TRUE, it returns NA values also.

Details

This will replace IndiNCA.

Value

CMAX

maximum concentration, Cmax

CMAXD

dose normalized Cmax, CMAX / Dose, Cmax / Dose

TMAX

time of maximum concentration, Tmax

TLAG

time to observe the first non-zero concentration, for extravascular administration only

CLST

last positive concentration observed, Clast

CLSTP

last positive concentration predicted, Clast_pred

TLST

time of last positive concentration, Tlast

LAMZHL

half-life by lambda z, ln(2)/LAMZ

LAMZ

lambda_z negative of best fit terminal slope

LAMZLL

earliest time for LAMZ

LAMZUL

last time for LAMZ

LAMZNPT

number of points for LAMZ

CORRXY

correlation of log(concentration) and time

R2

R-squared

R2ADJ

R-squared adjusted

C0

back extrapolated concentration at time 0, for bolus intravascular administration only

AUCLST

AUC from 0 to TLST

AUCALL

AUC using all the given points, including trailing zero concentrations

AUCIFO

AUC infinity observed

AUCIFOD

AUCIFO / Dose

AUCIFP

AUC infinity predicted using CLSTP instead of CLST

AUCIFPD

AUCIFP / Dose

AUCPEO

AUC % extrapolation observed

AUCPEP

AUC % extrapolated for AUCIFP

AUCPBEO

AUC % back extrapolation observed, for bolus IV administration only

AUCPBEP

AUC % back extrapolation predicted with AUCIFP, for bolus IV administration only

AUMCLST

AUMC to the TLST

AUMCIFO

AUMC infinity observed using CLST

AUMCIFP

AUMC infinity determined by CLSTP

AUMCPEO

AUMC % extrapolated observed

AUMCPEP

AUMC % extrapolated predicted

MRTIVLST

mean residence time (MRT) to TLST, for intravascular administration

MRTIVIFO

mean residence time (MRT) infinity using CLST, for intravascular administration

MRTIVIFP

mean residence time (MRT) infinity using CLSTP, for intravascular administration

MRTEVLST

mean residence time (MRT) to TLST, for extravascular administration

MRTEVIFO

mean residence time (MRT) infinity using CLST, for extravascular administration

MRTEVIFP

mean residence time (MRT) infinity using CLSTP, for extravascular administration

VZO

volume of distribution determined by LAMZ and AUCIFO, for intravascular administration

VZP

volume of distribution determined by LAMZ and AUCIFP, for intravascular administration

VZFO

VZO for extravascular administration, VZO/F, F is bioavailability

VZFP

VZP for extravascular administration, VZP/F, F is bioavailability

CLO

clearance using AUCIFO, for intravascular administration

CLP

clearance using AUCIFP, for intravascular administration

CLFO

CLO for extravascular administration, CLO/F, F is bioavailability

CLFP

CLP for extravascular administration, CLP/F, F is bioavailability

VSSO

volume of distribution at steady state using CLST, for intravascular administration only

VSSP

volume of distribution at steady state using CLSTP, for intravascular administration only

References

Gabrielsson J, Weiner D. Pharmacokinetic and Pharmacodynamic Data Analysis - Concepts and Applications. 5th ed. 2016.

See also

Examples

# For one subject
x = Theoph[Theoph$Subject=="1","Time"]
y = Theoph[Theoph$Subject=="1","conc"]

sNCA(x, y, dose=320, doseUnit="mg", concUnit="mg/L", timeUnit="h")

#>           b0         CMAX        CMAXD         TMAX         TLAG         CLST 
#>    2.3687851   10.5000000    0.0328125    1.1200000    0.0000000    3.2800000 
#>        CLSTP         TLST       LAMZHL         LAMZ       LAMZLL       LAMZUL 
#>    3.2801465   24.3700000   14.3043776    0.0484570    9.0500000   24.3700000 
#>      LAMZNPT       CORRXY           R2        R2ADJ           C0       AUCLST 
#>    3.0000000   -0.9999999    0.9999997    0.9999995           NA  148.9230500 
#>       AUCALL       AUCIFO      AUCIFOD       AUCIFP      AUCIFPD       AUCPEO 
#>  148.9230500  216.6119330    0.6769123  216.6149558    0.6769217   31.2489169 
#>       AUCPEP      AUCPBEO      AUCPBEP      AUMCLST      AUMCIFO      AUMCIFP 
#>   31.2498763           NA           NA 1459.0711035 4505.5348194 4505.6708646 
#>      AUMCPEO      AUMCPEP     MRTIVLST     MRTIVIFO     MRTIVIFP     MRTEVLST 
#>   67.6160287   67.6170065           NA           NA           NA    9.7974834 
#>     MRTEVIFO     MRTEVIFP          VZO          VZP         VZFO         VZFP 
#>   20.8000305   20.8003683           NA           NA   30.4867482   30.4863228 
#>          CLO          CLP         CLFO         CLFP         VSSO         VSSP 
#>           NA           NA    1.4772963    1.4772757           NA           NA 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        

sNCA(x, y, dose=320, concUnit="mg/L", returnNA=FALSE)

#>           b0         CMAX        CMAXD         TMAX         TLAG         CLST 
#>    2.3687851   10.5000000    0.0328125    1.1200000    0.0000000    3.2800000 
#>        CLSTP         TLST       LAMZHL         LAMZ       LAMZLL       LAMZUL 
#>    3.2801465   24.3700000   14.3043776    0.0484570    9.0500000   24.3700000 
#>      LAMZNPT       CORRXY           R2        R2ADJ       AUCLST       AUCALL 
#>    3.0000000   -0.9999999    0.9999997    0.9999995  148.9230500  148.9230500 
#>       AUCIFO      AUCIFOD       AUCIFP      AUCIFPD       AUCPEO       AUCPEP 
#>  216.6119330    0.6769123  216.6149558    0.6769217   31.2489169   31.2498763 
#>      AUMCLST      AUMCIFO      AUMCIFP      AUMCPEO      AUMCPEP     MRTEVLST 
#> 1459.0711035 4505.5348194 4505.6708646   67.6160287   67.6170065    9.7974834 
#>     MRTEVIFO     MRTEVIFP         VZFO         VZFP         CLFO         CLFP 
#>   20.8000305   20.8003683   30.4867482   30.4863228    1.4772963    1.4772757 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "h*mg/L"    "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"         "%"        
#> [25] "h2*mg/L"   "h2*mg/L"   "h2*mg/L"   "%"         "%"         "h"        
#> [31] "h"         "h"         "L"         "L"         "L/h"       "L/h"      

iAUC = data.frame(Name=c("AUC[0-12h]","AUC[0-24h]"), Start=c(0,0), End=c(12,24))
sNCA(x, y, dose=320, doseUnit="mg", concUnit="mg/L", timeUnit="h", iAUC=iAUC)

#>           b0         CMAX        CMAXD         TMAX         TLAG         CLST 
#>    2.3687851   10.5000000    0.0328125    1.1200000    0.0000000    3.2800000 
#>        CLSTP         TLST       LAMZHL         LAMZ       LAMZLL       LAMZUL 
#>    3.2801465   24.3700000   14.3043776    0.0484570    9.0500000   24.3700000 
#>      LAMZNPT       CORRXY           R2        R2ADJ           C0       AUCLST 
#>    3.0000000   -0.9999999    0.9999997    0.9999995           NA  148.9230500 
#>       AUCALL       AUCIFO      AUCIFOD       AUCIFP      AUCIFPD       AUCPEO 
#>  148.9230500  216.6119330    0.6769123  216.6149558    0.6769217   31.2489169 
#>       AUCPEP      AUCPBEO      AUCPBEP      AUMCLST      AUMCIFO      AUMCIFP 
#>   31.2498763           NA           NA 1459.0711035 4505.5348194 4505.6708646 
#>      AUMCPEO      AUMCPEP     MRTIVLST     MRTIVIFO     MRTIVIFP     MRTEVLST 
#>   67.6160287   67.6170065           NA           NA           NA    9.7974834 
#>     MRTEVIFO     MRTEVIFP          VZO          VZP         VZFO         VZFP 
#>   20.8000305   20.8003683           NA           NA   30.4867482   30.4863228 
#>          CLO          CLP         CLFO         CLFP         VSSO         VSSP 
#>           NA           NA    1.4772963    1.4772757           NA           NA 
#>   AUC[0-12h]   AUC[0-24h] 
#>   91.7355220  147.6945866 
#> attr(,"units")
#>  [1] ""          "mg/L"      "mg/L/mg"   "h"         "h"         "mg/L"     
#>  [7] "mg/L"      "h"         "h"         "/h"        "h"         "h"        
#> [13] ""          ""          ""          ""          "mg/L"      "h*mg/L"   
#> [19] "h*mg/L"    "h*mg/L"    "h*mg/L/mg" "h*mg/L"    "h*mg/L/mg" "%"        
#> [25] "%"         "%"         "%"         "h2*mg/L"   "h2*mg/L"   "h2*mg/L"  
#> [31] "%"         "%"         "h"         "h"         "h"         "h"        
#> [37] "h"         "h"         "L"         "L"         "L"         "L"        
#> [43] "L/h"       "L/h"       "L/h"       "L/h"       "L"         "L"        
#> [49] "h*mg/L"    "h*mg/L"   

MW = 180.164 # Molecular weight of theophylline

sNCA(x, y/MW, dose=320, doseUnit="mg", concUnit="mmol/L", timeUnit="h")

#> Warning: Molecular weight should be given for more informative results!

#>            b0          CMAX         CMAXD          TMAX          TLAG 
#> -2.8250824530  0.0582802336  0.0001821257  1.1200000000  0.0000000000 
#>          CLST         CLSTP          TLST        LAMZHL          LAMZ 
#>  0.0182056349  0.0182064479 24.3700000000 14.3043775711  0.0484569970 
#>        LAMZLL        LAMZUL       LAMZNPT        CORRXY            R2 
#>  9.0500000000 24.3700000000  3.0000000000 -0.9999998648  0.9999997297 
#>         R2ADJ            C0        AUCLST        AUCALL        AUCIFO 
#>  0.9999994593            NA  0.8265971559  0.8265971559  1.2023041953 
#>       AUCIFOD        AUCIFP       AUCIFPD        AUCPEO        AUCPEP 
#>  0.0037572006  1.2023209731  0.0037572530 31.2489169405 31.2498763313 
#>       AUCPBEO       AUCPBEP       AUMCLST       AUMCIFO       AUMCIFP 
#>            NA            NA  8.0985718762 25.0079639629 25.0087190814 
#>       AUMCPEO       AUMCPEP      MRTIVLST      MRTIVIFO      MRTIVIFP 
#> 67.6160286851 67.6170064935            NA            NA            NA 
#>      MRTEVLST      MRTEVIFO      MRTEVIFP           VZO           VZP 
#>  9.7974833547 20.8000305256 20.8003683211            NA            NA 
#>          VZFO          VZFP           CLO           CLP          CLFO 
#>            NA            NA            NA            NA            NA 
#>          CLFP          VSSO          VSSP 
#>            NA            NA            NA 
#> attr(,"units")
#>  [1] ""            "mmol/L"      "mmol/L/mg"   "h"           "h"          
#>  [6] "mmol/L"      "mmol/L"      "h"           "h"           "/h"         
#> [11] "h"           "h"           ""            ""            ""           
#> [16] ""            "mmol/L"      "h*mmol/L"    "h*mmol/L"    "h*mmol/L"   
#> [21] "h*mmol/L/mg" "h*mmol/L"    "h*mmol/L/mg" "%"           "%"          
#> [26] "%"           "%"           "h2*mmol/L"   "h2*mmol/L"   "h2*mmol/L"  
#> [31] "%"           "%"           "h"           "h"           "h"          
#> [36] "h"           "h"           "h"           "L"           "L"          
#> [41] "L"           "L"           "L/h"         "L/h"         "L/h"        
#> [46] "L/h"         "L"           "L"          

sNCA(x, y/MW, dose=320, doseUnit="mg", concUnit="mmol/L", timeUnit="h", MW=MW)

#>            b0          CMAX         CMAXD          TMAX          TLAG 
#> -2.8250824530  0.0582802336  0.0001821257  1.1200000000  0.0000000000 
#>          CLST         CLSTP          TLST        LAMZHL          LAMZ 
#>  0.0182056349  0.0182064479 24.3700000000 14.3043775711  0.0484569970 
#>        LAMZLL        LAMZUL       LAMZNPT        CORRXY            R2 
#>  9.0500000000 24.3700000000  3.0000000000 -0.9999998648  0.9999997297 
#>         R2ADJ            C0        AUCLST        AUCALL        AUCIFO 
#>  0.9999994593            NA  0.8265971559  0.8265971559  1.2023041953 
#>       AUCIFOD        AUCIFP       AUCIFPD        AUCPEO        AUCPEP 
#>  0.0037572006  1.2023209731  0.0037572530 31.2489169405 31.2498763313 
#>       AUCPBEO       AUCPBEP       AUMCLST       AUMCIFO       AUMCIFP 
#>            NA            NA  8.0985718762 25.0079639629 25.0087190814 
#>       AUMCPEO       AUMCPEP      MRTIVLST      MRTIVIFO      MRTIVIFP 
#> 67.6160286851 67.6170064935            NA            NA            NA 
#>      MRTEVLST      MRTEVIFO      MRTEVIFP           VZO           VZP 
#>  9.7974833547 20.8000305256 20.8003683211            NA            NA 
#>          VZFO          VZFP           CLO           CLP          CLFO 
#> 30.4867482346 30.4863228055            NA            NA  1.4772962667 
#>          CLFP          VSSO          VSSP 
#>  1.4772756517            NA            NA 
#> attr(,"units")
#>  [1] ""            "mmol/L"      "mmol/L/mg"   "h"           "h"          
#>  [6] "mmol/L"      "mmol/L"      "h"           "h"           "/h"         
#> [11] "h"           "h"           ""            ""            ""           
#> [16] ""            "mmol/L"      "h*mmol/L"    "h*mmol/L"    "h*mmol/L"   
#> [21] "h*mmol/L/mg" "h*mmol/L"    "h*mmol/L/mg" "%"           "%"          
#> [26] "%"           "%"           "h2*mmol/L"   "h2*mmol/L"   "h2*mmol/L"  
#> [31] "%"           "%"           "h"           "h"           "h"          
#> [36] "h"           "h"           "h"           "L"           "L"          
#> [41] "L"           "L"           "L/h"         "L/h"         "L/h"        
#> [46] "L/h"         "L"           "L"          

sNCA(x, y, dose=320/MW, doseUnit="mmol", concUnit="mg/L", timeUnit="h", MW=MW)

#>           b0         CMAX        CMAXD         TMAX         TLAG         CLST 
#>    2.3687851   10.5000000    5.9116312    1.1200000    0.0000000    3.2800000 
#>        CLSTP         TLST       LAMZHL         LAMZ       LAMZLL       LAMZUL 
#>    3.2801465   24.3700000   14.3043776    0.0484570    9.0500000   24.3700000 
#>      LAMZNPT       CORRXY           R2        R2ADJ           C0       AUCLST 
#>    3.0000000   -0.9999999    0.9999997    0.9999995           NA  148.9230500 
#>       AUCALL       AUCIFO      AUCIFOD       AUCIFP      AUCIFPD       AUCPEO 
#>  148.9230500  216.6119330  121.9552259  216.6149558  121.9569278   31.2489169 
#>       AUCPEP      AUCPBEO      AUCPBEP      AUMCLST      AUMCIFO      AUMCIFP 
#>   31.2498763           NA           NA 1459.0711035 4505.5348194 4505.6708646 
#>      AUMCPEO      AUMCPEP     MRTIVLST     MRTIVIFO     MRTIVIFP     MRTEVLST 
#>   67.6160287   67.6170065           NA           NA           NA    9.7974834 
#>     MRTEVIFO     MRTEVIFP          VZO          VZP         VZFO         VZFP 
#>   20.8000305   20.8003683           NA           NA   30.4867482   30.4863228 
#>          CLO          CLP         CLFO         CLFP         VSSO         VSSP 
#>           NA           NA    1.4772963    1.4772757           NA           NA 
#> attr(,"units")
#>  [1] ""            "mg/L"        "mg/L/mmol"   "h"           "h"          
#>  [6] "mg/L"        "mg/L"        "h"           "h"           "/h"         
#> [11] "h"           "h"           ""            ""            ""           
#> [16] ""            "mg/L"        "h*mg/L"      "h*mg/L"      "h*mg/L"     
#> [21] "h*mg/L/mmol" "h*mg/L"      "h*mg/L/mmol" "%"           "%"          
#> [26] "%"           "%"           "h2*mg/L"     "h2*mg/L"     "h2*mg/L"    
#> [31] "%"           "%"           "h"           "h"           "h"          
#> [36] "h"           "h"           "h"           "L"           "L"          
#> [41] "L"           "L"           "L/h"         "L/h"         "L/h"        
#> [46] "L/h"         "L"           "L"          

sNCA(x, y/MW, dose=320/MW, doseUnit="mmol", concUnit="mmol/L", timeUnit="h", MW=MW)

#>          b0        CMAX       CMAXD        TMAX        TLAG        CLST 
#> -2.82508245  0.05828023  0.03281250  1.12000000  0.00000000  0.01820563 
#>       CLSTP        TLST      LAMZHL        LAMZ      LAMZLL      LAMZUL 
#>  0.01820645 24.37000000 14.30437757  0.04845700  9.05000000 24.37000000 
#>     LAMZNPT      CORRXY          R2       R2ADJ          C0      AUCLST 
#>  3.00000000 -0.99999986  0.99999973  0.99999946          NA  0.82659716 
#>      AUCALL      AUCIFO     AUCIFOD      AUCIFP     AUCIFPD      AUCPEO 
#>  0.82659716  1.20230420  0.67691229  1.20232097  0.67692174 31.24891694 
#>      AUCPEP     AUCPBEO     AUCPBEP     AUMCLST     AUMCIFO     AUMCIFP 
#> 31.24987633          NA          NA  8.09857188 25.00796396 25.00871908 
#>     AUMCPEO     AUMCPEP    MRTIVLST    MRTIVIFO    MRTIVIFP    MRTEVLST 
#> 67.61602869 67.61700649          NA          NA          NA  9.79748335 
#>    MRTEVIFO    MRTEVIFP         VZO         VZP        VZFO        VZFP 
#> 20.80003053 20.80036832          NA          NA 30.48674823 30.48632281 
#>         CLO         CLP        CLFO        CLFP        VSSO        VSSP 
#>          NA          NA  1.47729627  1.47727565          NA          NA 
#> attr(,"units")
#>  [1] ""              "mmol/L"        "mmol/L/mmol"   "h"            
#>  [5] "h"             "mmol/L"        "mmol/L"        "h"            
#>  [9] "h"             "/h"            "h"             "h"            
#> [13] ""              ""              ""              ""             
#> [17] "mmol/L"        "h*mmol/L"      "h*mmol/L"      "h*mmol/L"     
#> [21] "h*mmol/L/mmol" "h*mmol/L"      "h*mmol/L/mmol" "%"            
#> [25] "%"             "%"             "%"             "h2*mmol/L"    
#> [29] "h2*mmol/L"     "h2*mmol/L"     "%"             "%"            
#> [33] "h"             "h"             "h"             "h"            
#> [37] "h"             "h"             "L"             "L"            
#> [41] "L"             "L"             "L/h"           "L/h"          
#> [45] "L/h"           "L/h"           "L"             "L"            

sNCA(x, y/MW, dose=320/MW, doseUnit="mmol", concUnit="mmol/L", timeUnit="h", MW=MW,
     returnNA=FALSE)

#>          b0        CMAX       CMAXD        TMAX        TLAG        CLST 
#> -2.82508245  0.05828023  0.03281250  1.12000000  0.00000000  0.01820563 
#>       CLSTP        TLST      LAMZHL        LAMZ      LAMZLL      LAMZUL 
#>  0.01820645 24.37000000 14.30437757  0.04845700  9.05000000 24.37000000 
#>     LAMZNPT      CORRXY          R2       R2ADJ      AUCLST      AUCALL 
#>  3.00000000 -0.99999986  0.99999973  0.99999946  0.82659716  0.82659716 
#>      AUCIFO     AUCIFOD      AUCIFP     AUCIFPD      AUCPEO      AUCPEP 
#>  1.20230420  0.67691229  1.20232097  0.67692174 31.24891694 31.24987633 
#>     AUMCLST     AUMCIFO     AUMCIFP     AUMCPEO     AUMCPEP    MRTEVLST 
#>  8.09857188 25.00796396 25.00871908 67.61602869 67.61700649  9.79748335 
#>    MRTEVIFO    MRTEVIFP        VZFO        VZFP        CLFO        CLFP 
#> 20.80003053 20.80036832 30.48674823 30.48632281  1.47729627  1.47727565 
#> attr(,"units")
#>  [1] ""              "mmol/L"        "mmol/L/mmol"   "h"            
#>  [5] "h"             "mmol/L"        "mmol/L"        "h"            
#>  [9] "h"             "/h"            "h"             "h"            
#> [13] ""              ""              ""              ""             
#> [17] "h*mmol/L"      "h*mmol/L"      "h*mmol/L"      "h*mmol/L/mmol"
#> [21] "h*mmol/L"      "h*mmol/L/mmol" "%"             "%"            
#> [25] "h2*mmol/L"     "h2*mmol/L"     "h2*mmol/L"     "%"            
#> [29] "%"             "h"             "h"             "h"            
#> [33] "L"             "L"             "L/h"           "L/h"          

sNCA(x, y/MW, doseUnit="mmol", concUnit="mmol/L", timeUnit="h", MW=MW, returnNA=FALSE)

#>          b0        CMAX        TMAX        TLAG        CLST       CLSTP 
#> -2.82508245  0.05828023  1.12000000  0.00000000  0.01820563  0.01820645 
#>        TLST      LAMZHL        LAMZ      LAMZLL      LAMZUL     LAMZNPT 
#> 24.37000000 14.30437757  0.04845700  9.05000000 24.37000000  3.00000000 
#>      CORRXY          R2       R2ADJ      AUCLST      AUCALL      AUCIFO 
#> -0.99999986  0.99999973  0.99999946  0.82659716  0.82659716  1.20230420 
#>      AUCIFP      AUCPEO      AUCPEP     AUMCLST     AUMCIFO     AUMCIFP 
#>  1.20232097 31.24891694 31.24987633  8.09857188 25.00796396 25.00871908 
#>     AUMCPEO     AUMCPEP    MRTEVLST    MRTEVIFO    MRTEVIFP        VZFO 
#> 67.61602869 67.61700649  9.79748335 20.80003053 20.80036832  0.00000000 
#>        VZFP        CLFO        CLFP 
#>  0.00000000  0.00000000  0.00000000 
#> attr(,"units")
#>  [1] ""          "mmol/L"    "h"         "h"         "mmol/L"    "mmol/L"   
#>  [7] "h"         "h"         "/h"        "h"         "h"         ""         
#> [13] ""          ""          ""          "h*mmol/L"  "h*mmol/L"  "h*mmol/L" 
#> [19] "h*mmol/L"  "%"         "%"         "h2*mmol/L" "h2*mmol/L" "h2*mmol/L"
#> [25] "%"         "%"         "h"         "h"         "h"         "L"        
#> [31] "L"         "L/h"       "L/h"      

sNCA(x, y/MW, dose=as.numeric(NA), doseUnit="mmol", concUnit="mmol/L", timeUnit="h",
     MW=MW, returnNA=FALSE)

#>          b0        CMAX        TMAX        TLAG        CLST       CLSTP 
#> -2.82508245  0.05828023  1.12000000  0.00000000  0.01820563  0.01820645 
#>        TLST      LAMZHL        LAMZ      LAMZLL      LAMZUL     LAMZNPT 
#> 24.37000000 14.30437757  0.04845700  9.05000000 24.37000000  3.00000000 
#>      CORRXY          R2       R2ADJ      AUCLST      AUCALL      AUCIFO 
#> -0.99999986  0.99999973  0.99999946  0.82659716  0.82659716  1.20230420 
#>      AUCIFP      AUCPEO      AUCPEP     AUMCLST     AUMCIFO     AUMCIFP 
#>  1.20232097 31.24891694 31.24987633  8.09857188 25.00796396 25.00871908 
#>     AUMCPEO     AUMCPEP    MRTEVLST    MRTEVIFO    MRTEVIFP 
#> 67.61602869 67.61700649  9.79748335 20.80003053 20.80036832 
#> attr(,"units")
#>  [1] ""          "mmol/L"    "h"         "h"         "mmol/L"    "mmol/L"   
#>  [7] "h"         "h"         "/h"        "h"         "h"         ""         
#> [13] ""          ""          ""          "h*mmol/L"  "h*mmol/L"  "h*mmol/L" 
#> [19] "h*mmol/L"  "%"         "%"         "h2*mmol/L" "h2*mmol/L" "h2*mmol/L"
#> [25] "%"         "%"         "h"         "h"         "h"        

sNCA(x, y, dose=320, concUnit="mg/L", timeUnit="hr")

#>           b0         CMAX        CMAXD         TMAX         TLAG         CLST 
#>    2.3687851   10.5000000    0.0328125    1.1200000    0.0000000    3.2800000 
#>        CLSTP         TLST       LAMZHL         LAMZ       LAMZLL       LAMZUL 
#>    3.2801465   24.3700000   14.3043776    0.0484570    9.0500000   24.3700000 
#>      LAMZNPT       CORRXY           R2        R2ADJ           C0       AUCLST 
#>    3.0000000   -0.9999999    0.9999997    0.9999995           NA  148.9230500 
#>       AUCALL       AUCIFO      AUCIFOD       AUCIFP      AUCIFPD       AUCPEO 
#>  148.9230500  216.6119330    0.6769123  216.6149558    0.6769217   31.2489169 
#>       AUCPEP      AUCPBEO      AUCPBEP      AUMCLST      AUMCIFO      AUMCIFP 
#>   31.2498763           NA           NA 1459.0711035 4505.5348194 4505.6708646 
#>      AUMCPEO      AUMCPEP     MRTIVLST     MRTIVIFO     MRTIVIFP     MRTEVLST 
#>   67.6160287   67.6170065           NA           NA           NA    9.7974834 
#>     MRTEVIFO     MRTEVIFP          VZO          VZP         VZFO         VZFP 
#>   20.8000305   20.8003683           NA           NA   30.4867482   30.4863228 
#>          CLO          CLP         CLFO         CLFP         VSSO         VSSP 
#>           NA           NA    1.4772963    1.4772757           NA           NA 
#> attr(,"units")
#>  [1] ""           "mg/L"       "mg/L/mg"    "hr"         "hr"        
#>  [6] "mg/L"       "mg/L"       "hr"         "hr"         "/hr"       
#> [11] "hr"         "hr"         ""           ""           ""          
#> [16] ""           "mg/L"       "hr*mg/L"    "hr*mg/L"    "hr*mg/L"   
#> [21] "hr*mg/L/mg" "hr*mg/L"    "hr*mg/L/mg" "%"          "%"         
#> [26] "%"          "%"          "hr2*mg/L"   "hr2*mg/L"   "hr2*mg/L"  
#> [31] "%"          "%"          "hr"         "hr"         "hr"        
#> [36] "hr"         "hr"         "hr"         "L"          "L"         
#> [41] "L"          "L"          "L/hr"       "L/hr"       "L/hr"      
#> [46] "L/hr"       "L"          "L"         

sNCA(x*60, y, dose=320, concUnit="mg/L", timeUnit="min")

#>            b0          CMAX         CMAXD          TMAX          TLAG 
#>  2.368785e+00  1.050000e+01  3.281250e-02  6.720000e+01  0.000000e+00 
#>          CLST         CLSTP          TLST        LAMZHL          LAMZ 
#>  3.280000e+00  3.280146e+00  1.462200e+03  8.582627e+02  8.076166e-04 
#>        LAMZLL        LAMZUL       LAMZNPT        CORRXY            R2 
#>  5.430000e+02  1.462200e+03  3.000000e+00 -9.999999e-01  9.999997e-01 
#>         R2ADJ            C0        AUCLST        AUCALL        AUCIFO 
#>  9.999995e-01            NA  8.935383e+03  8.935383e+03  1.299672e+04 
#>       AUCIFOD        AUCIFP       AUCIFPD        AUCPEO        AUCPEP 
#>  4.061474e+01  1.299690e+04  4.061530e+01  3.124892e+01  3.124988e+01 
#>       AUCPBEO       AUCPBEP       AUMCLST       AUMCIFO       AUMCIFP 
#>            NA            NA  5.252656e+06  1.621993e+07  1.622042e+07 
#>       AUMCPEO       AUMCPEP      MRTIVLST      MRTIVIFO      MRTIVIFP 
#>  6.761603e+01  6.761701e+01            NA            NA            NA 
#>      MRTEVLST      MRTEVIFO      MRTEVIFP           VZO           VZP 
#>  5.878490e+02  1.248002e+03  1.248022e+03            NA            NA 
#>          VZFO          VZFP           CLO           CLP          CLFO 
#>  3.048675e+01  3.048632e+01            NA            NA  2.462160e-02 
#>          CLFP          VSSO          VSSP 
#>  2.462126e-02            NA            NA 
#> attr(,"units")
#>  [1] ""            "mg/L"        "mg/L/mg"     "min"         "min"        
#>  [6] "mg/L"        "mg/L"        "min"         "min"         "/min"       
#> [11] "min"         "min"         ""            ""            ""           
#> [16] ""            "mg/L"        "min*mg/L"    "min*mg/L"    "min*mg/L"   
#> [21] "min*mg/L/mg" "min*mg/L"    "min*mg/L/mg" "%"           "%"          
#> [26] "%"           "%"           "min2*mg/L"   "min2*mg/L"   "min2*mg/L"  
#> [31] "%"           "%"           "min"         "min"         "min"        
#> [36] "min"         "min"         "min"         "L"           "L"          
#> [41] "L"           "L"           "L/min"       "L/min"       "L/min"      
#> [46] "L/min"       "L"           "L"