Disply CDISC standard units and multiplied factor of NCA results

It displays CDISC PP output units and multiplication factor for them.

Unit(code = "", timeUnit = "h", concUnit = "ng/mL", doseUnit = "mg", MW = 0)

Arguments

code
vector of PPTESTCD

timeUnit
unit of time

concUnit
unit of concentration

doseUnit
unit of dose

MW
molecular weight of drug

Value

row names

PPTESTCD

Unit

unit

Factor

internal mulitpilcation factor

Examples

Unit(concUnit="ug/L", doseUnit="mg")

#>               Unit Factor
#> b0                      1
#> CMAX          ug/L      1
#> CMAXD      ug/L/mg      1
#> TMAX             h      1
#> TLAG             h      1
#> CLST          ug/L      1
#> CLSTP         ug/L      1
#> TLST             h      1
#> LAMZHL           h      1
#> LAMZ            /h      1
#> LAMZLL           h      1
#> LAMZUL           h      1
#> LAMZNPT                 1
#> CORRXY                  1
#> R2                      1
#> R2ADJ                   1
#> C0            ug/L      1
#> AUCLST      h*ug/L      1
#> AUCALL      h*ug/L      1
#> AUCIFO      h*ug/L      1
#> AUCIFOD  h*ug/L/mg      1
#> AUCIFP      h*ug/L      1
#> AUCIFPD  h*ug/L/mg      1
#> AUCPEO           %      1
#> AUCPEP           %      1
#> AUCPBEO          %      1
#> AUCPBEP          %      1
#> AUMCLST    h2*ug/L      1
#> AUMCIFO    h2*ug/L      1
#> AUMCIFP    h2*ug/L      1
#> AUMCPEO          %      1
#> AUMCPEP          %      1
#> MRTIVLST         h      1
#> MRTIVIFO         h      1
#> MRTIVIFP         h      1
#> MRTEVLST         h      1
#> MRTEVIFO         h      1
#> MRTEVIFP         h      1
#> VZO              L   1000
#> VZP              L   1000
#> VZFO             L   1000
#> VZFP             L   1000
#> CLO            L/h   1000
#> CLP            L/h   1000
#> CLFO           L/h   1000
#> CLFP           L/h   1000
#> VSSO             L   1000
#> VSSP             L   1000

Unit(concUnit="ng/L", doseUnit="mg")

#>               Unit Factor
#> b0                  1e+00
#> CMAX          ng/L  1e+00
#> CMAXD      ng/L/mg  1e+00
#> TMAX             h  1e+00
#> TLAG             h  1e+00
#> CLST          ng/L  1e+00
#> CLSTP         ng/L  1e+00
#> TLST             h  1e+00
#> LAMZHL           h  1e+00
#> LAMZ            /h  1e+00
#> LAMZLL           h  1e+00
#> LAMZUL           h  1e+00
#> LAMZNPT             1e+00
#> CORRXY              1e+00
#> R2                  1e+00
#> R2ADJ               1e+00
#> C0            ng/L  1e+00
#> AUCLST      h*ng/L  1e+00
#> AUCALL      h*ng/L  1e+00
#> AUCIFO      h*ng/L  1e+00
#> AUCIFOD  h*ng/L/mg  1e+00
#> AUCIFP      h*ng/L  1e+00
#> AUCIFPD  h*ng/L/mg  1e+00
#> AUCPEO           %  1e+00
#> AUCPEP           %  1e+00
#> AUCPBEO          %  1e+00
#> AUCPBEP          %  1e+00
#> AUMCLST    h2*ng/L  1e+00
#> AUMCIFO    h2*ng/L  1e+00
#> AUMCIFP    h2*ng/L  1e+00
#> AUMCPEO          %  1e+00
#> AUMCPEP          %  1e+00
#> MRTIVLST         h  1e+00
#> MRTIVIFO         h  1e+00
#> MRTIVIFP         h  1e+00
#> MRTEVLST         h  1e+00
#> MRTEVIFO         h  1e+00
#> MRTEVIFP         h  1e+00
#> VZO              L  1e+06
#> VZP              L  1e+06
#> VZFO             L  1e+06
#> VZFP             L  1e+06
#> CLO            L/h  1e+06
#> CLP            L/h  1e+06
#> CLFO           L/h  1e+06
#> CLFP           L/h  1e+06
#> VSSO             L  1e+06
#> VSSP             L  1e+06

Unit(concUnit="umol/L", doseUnit="mmol")

#>                   Unit Factor
#> b0                          1
#> CMAX            umol/L      1
#> CMAXD      umol/L/mmol      1
#> TMAX                 h      1
#> TLAG                 h      1
#> CLST            umol/L      1
#> CLSTP           umol/L      1
#> TLST                 h      1
#> LAMZHL               h      1
#> LAMZ                /h      1
#> LAMZLL               h      1
#> LAMZUL               h      1
#> LAMZNPT                     1
#> CORRXY                      1
#> R2                          1
#> R2ADJ                       1
#> C0              umol/L      1
#> AUCLST        h*umol/L      1
#> AUCALL        h*umol/L      1
#> AUCIFO        h*umol/L      1
#> AUCIFOD  h*umol/L/mmol      1
#> AUCIFP        h*umol/L      1
#> AUCIFPD  h*umol/L/mmol      1
#> AUCPEO               %      1
#> AUCPEP               %      1
#> AUCPBEO              %      1
#> AUCPBEP              %      1
#> AUMCLST      h2*umol/L      1
#> AUMCIFO      h2*umol/L      1
#> AUMCIFP      h2*umol/L      1
#> AUMCPEO              %      1
#> AUMCPEP              %      1
#> MRTIVLST             h      1
#> MRTIVIFO             h      1
#> MRTIVIFP             h      1
#> MRTEVLST             h      1
#> MRTEVIFO             h      1
#> MRTEVIFP             h      1
#> VZO                  L   1000
#> VZP                  L   1000
#> VZFO                 L   1000
#> VZFP                 L   1000
#> CLO                L/h   1000
#> CLP                L/h   1000
#> CLFO               L/h   1000
#> CLFP               L/h   1000
#> VSSO                 L   1000
#> VSSP                 L   1000

Unit(concUnit="nmol/L", doseUnit="mmol")

#>                   Unit Factor
#> b0                      1e+00
#> CMAX            nmol/L  1e+00
#> CMAXD      nmol/L/mmol  1e+00
#> TMAX                 h  1e+00
#> TLAG                 h  1e+00
#> CLST            nmol/L  1e+00
#> CLSTP           nmol/L  1e+00
#> TLST                 h  1e+00
#> LAMZHL               h  1e+00
#> LAMZ                /h  1e+00
#> LAMZLL               h  1e+00
#> LAMZUL               h  1e+00
#> LAMZNPT                 1e+00
#> CORRXY                  1e+00
#> R2                      1e+00
#> R2ADJ                   1e+00
#> C0              nmol/L  1e+00
#> AUCLST        h*nmol/L  1e+00
#> AUCALL        h*nmol/L  1e+00
#> AUCIFO        h*nmol/L  1e+00
#> AUCIFOD  h*nmol/L/mmol  1e+00
#> AUCIFP        h*nmol/L  1e+00
#> AUCIFPD  h*nmol/L/mmol  1e+00
#> AUCPEO               %  1e+00
#> AUCPEP               %  1e+00
#> AUCPBEO              %  1e+00
#> AUCPBEP              %  1e+00
#> AUMCLST      h2*nmol/L  1e+00
#> AUMCIFO      h2*nmol/L  1e+00
#> AUMCIFP      h2*nmol/L  1e+00
#> AUMCPEO              %  1e+00
#> AUMCPEP              %  1e+00
#> MRTIVLST             h  1e+00
#> MRTIVIFO             h  1e+00
#> MRTIVIFP             h  1e+00
#> MRTEVLST             h  1e+00
#> MRTEVIFO             h  1e+00
#> MRTEVIFP             h  1e+00
#> VZO                  L  1e+06
#> VZP                  L  1e+06
#> VZFO                 L  1e+06
#> VZFP                 L  1e+06
#> CLO                L/h  1e+06
#> CLP                L/h  1e+06
#> CLFO               L/h  1e+06
#> CLFP               L/h  1e+06
#> VSSO                 L  1e+06
#> VSSP                 L  1e+06

Unit(concUnit="mmol/L", doseUnit="mg", MW=500)

#>                 Unit Factor
#> b0                    1.000
#> CMAX          mmol/L  1.000
#> CMAXD      mmol/L/mg  1.000
#> TMAX               h  1.000
#> TLAG               h  1.000
#> CLST          mmol/L  1.000
#> CLSTP         mmol/L  1.000
#> TLST               h  1.000
#> LAMZHL             h  1.000
#> LAMZ              /h  1.000
#> LAMZLL             h  1.000
#> LAMZUL             h  1.000
#> LAMZNPT               1.000
#> CORRXY                1.000
#> R2                    1.000
#> R2ADJ                 1.000
#> C0            mmol/L  1.000
#> AUCLST      h*mmol/L  1.000
#> AUCALL      h*mmol/L  1.000
#> AUCIFO      h*mmol/L  1.000
#> AUCIFOD  h*mmol/L/mg  1.000
#> AUCIFP      h*mmol/L  1.000
#> AUCIFPD  h*mmol/L/mg  1.000
#> AUCPEO             %  1.000
#> AUCPEP             %  1.000
#> AUCPBEO            %  1.000
#> AUCPBEP            %  1.000
#> AUMCLST    h2*mmol/L  1.000
#> AUMCIFO    h2*mmol/L  1.000
#> AUMCIFP    h2*mmol/L  1.000
#> AUMCPEO            %  1.000
#> AUMCPEP            %  1.000
#> MRTIVLST           h  1.000
#> MRTIVIFO           h  1.000
#> MRTIVIFP           h  1.000
#> MRTEVLST           h  1.000
#> MRTEVIFO           h  1.000
#> MRTEVIFP           h  1.000
#> VZO                L  0.002
#> VZP                L  0.002
#> VZFO               L  0.002
#> VZFP               L  0.002
#> CLO              L/h  0.002
#> CLP              L/h  0.002
#> CLFO             L/h  0.002
#> CLFP             L/h  0.002
#> VSSO               L  0.002
#> VSSP               L  0.002

Unit(concUnit="umol/L", doseUnit="mg", MW=500)

#>                 Unit Factor
#> b0                        1
#> CMAX          umol/L      1
#> CMAXD      umol/L/mg      1
#> TMAX               h      1
#> TLAG               h      1
#> CLST          umol/L      1
#> CLSTP         umol/L      1
#> TLST               h      1
#> LAMZHL             h      1
#> LAMZ              /h      1
#> LAMZLL             h      1
#> LAMZUL             h      1
#> LAMZNPT                   1
#> CORRXY                    1
#> R2                        1
#> R2ADJ                     1
#> C0            umol/L      1
#> AUCLST      h*umol/L      1
#> AUCALL      h*umol/L      1
#> AUCIFO      h*umol/L      1
#> AUCIFOD  h*umol/L/mg      1
#> AUCIFP      h*umol/L      1
#> AUCIFPD  h*umol/L/mg      1
#> AUCPEO             %      1
#> AUCPEP             %      1
#> AUCPBEO            %      1
#> AUCPBEP            %      1
#> AUMCLST    h2*umol/L      1
#> AUMCIFO    h2*umol/L      1
#> AUMCIFP    h2*umol/L      1
#> AUMCPEO            %      1
#> AUMCPEP            %      1
#> MRTIVLST           h      1
#> MRTIVIFO           h      1
#> MRTIVIFP           h      1
#> MRTEVLST           h      1
#> MRTEVIFO           h      1
#> MRTEVIFP           h      1
#> VZO                L      2
#> VZP                L      2
#> VZFO               L      2
#> VZFP               L      2
#> CLO              L/h      2
#> CLP              L/h      2
#> CLFO             L/h      2
#> CLFP             L/h      2
#> VSSO               L      2
#> VSSP               L      2

Unit(concUnit="nmol/L", doseUnit="mg", MW=500)

#>                 Unit Factor
#> b0                        1
#> CMAX          nmol/L      1
#> CMAXD      nmol/L/mg      1
#> TMAX               h      1
#> TLAG               h      1
#> CLST          nmol/L      1
#> CLSTP         nmol/L      1
#> TLST               h      1
#> LAMZHL             h      1
#> LAMZ              /h      1
#> LAMZLL             h      1
#> LAMZUL             h      1
#> LAMZNPT                   1
#> CORRXY                    1
#> R2                        1
#> R2ADJ                     1
#> C0            nmol/L      1
#> AUCLST      h*nmol/L      1
#> AUCALL      h*nmol/L      1
#> AUCIFO      h*nmol/L      1
#> AUCIFOD  h*nmol/L/mg      1
#> AUCIFP      h*nmol/L      1
#> AUCIFPD  h*nmol/L/mg      1
#> AUCPEO             %      1
#> AUCPEP             %      1
#> AUCPBEO            %      1
#> AUCPBEP            %      1
#> AUMCLST    h2*nmol/L      1
#> AUMCIFO    h2*nmol/L      1
#> AUMCIFP    h2*nmol/L      1
#> AUMCPEO            %      1
#> AUMCPEP            %      1
#> MRTIVLST           h      1
#> MRTIVIFO           h      1
#> MRTIVIFP           h      1
#> MRTEVLST           h      1
#> MRTEVIFO           h      1
#> MRTEVIFP           h      1
#> VZO                L   2000
#> VZP                L   2000
#> VZFO               L   2000
#> VZFP               L   2000
#> CLO              L/h   2000
#> CLP              L/h   2000
#> CLFO             L/h   2000
#> CLFP             L/h   2000
#> VSSO               L   2000
#> VSSP               L   2000

Unit(concUnit="nmol/mL", doseUnit="mg", MW=500)

#>                 Unit Factor
#> b0                        1
#> CMAX          umol/L      1
#> CMAXD      umol/L/mg      1
#> TMAX               h      1
#> TLAG               h      1
#> CLST          umol/L      1
#> CLSTP         umol/L      1
#> TLST               h      1
#> LAMZHL             h      1
#> LAMZ              /h      1
#> LAMZLL             h      1
#> LAMZUL             h      1
#> LAMZNPT                   1
#> CORRXY                    1
#> R2                        1
#> R2ADJ                     1
#> C0            umol/L      1
#> AUCLST      h*umol/L      1
#> AUCALL      h*umol/L      1
#> AUCIFO      h*umol/L      1
#> AUCIFOD  h*umol/L/mg      1
#> AUCIFP      h*umol/L      1
#> AUCIFPD  h*umol/L/mg      1
#> AUCPEO             %      1
#> AUCPEP             %      1
#> AUCPBEO            %      1
#> AUCPBEP            %      1
#> AUMCLST    h2*umol/L      1
#> AUMCIFO    h2*umol/L      1
#> AUMCIFP    h2*umol/L      1
#> AUMCPEO            %      1
#> AUMCPEP            %      1
#> MRTIVLST           h      1
#> MRTIVIFO           h      1
#> MRTIVIFP           h      1
#> MRTEVLST           h      1
#> MRTEVIFO           h      1
#> MRTEVIFP           h      1
#> VZO                L      2
#> VZP                L      2
#> VZFO               L      2
#> VZFP               L      2
#> CLO              L/h      2
#> CLP              L/h      2
#> CLFO             L/h      2
#> CLFP             L/h      2
#> VSSO               L      2
#> VSSP               L      2

Unit(concUnit="ug/L", doseUnit="mmol", MW=500)

#>                 Unit Factor
#> b0                    1e+00
#> CMAX            ug/L  1e+00
#> CMAXD      ug/L/mmol  1e+00
#> TMAX               h  1e+00
#> TLAG               h  1e+00
#> CLST            ug/L  1e+00
#> CLSTP           ug/L  1e+00
#> TLST               h  1e+00
#> LAMZHL             h  1e+00
#> LAMZ              /h  1e+00
#> LAMZLL             h  1e+00
#> LAMZUL             h  1e+00
#> LAMZNPT               1e+00
#> CORRXY                1e+00
#> R2                    1e+00
#> R2ADJ                 1e+00
#> C0              ug/L  1e+00
#> AUCLST        h*ug/L  1e+00
#> AUCALL        h*ug/L  1e+00
#> AUCIFO        h*ug/L  1e+00
#> AUCIFOD  h*ug/L/mmol  1e+00
#> AUCIFP        h*ug/L  1e+00
#> AUCIFPD  h*ug/L/mmol  1e+00
#> AUCPEO             %  1e+00
#> AUCPEP             %  1e+00
#> AUCPBEO            %  1e+00
#> AUCPBEP            %  1e+00
#> AUMCLST      h2*ug/L  1e+00
#> AUMCIFO      h2*ug/L  1e+00
#> AUMCIFP      h2*ug/L  1e+00
#> AUMCPEO            %  1e+00
#> AUMCPEP            %  1e+00
#> MRTIVLST           h  1e+00
#> MRTIVIFO           h  1e+00
#> MRTIVIFP           h  1e+00
#> MRTEVLST           h  1e+00
#> MRTEVIFO           h  1e+00
#> MRTEVIFP           h  1e+00
#> VZO                L  5e+05
#> VZP                L  5e+05
#> VZFO               L  5e+05
#> VZFP               L  5e+05
#> CLO              L/h  5e+05
#> CLP              L/h  5e+05
#> CLFO             L/h  5e+05
#> CLFP             L/h  5e+05
#> VSSO               L  5e+05
#> VSSP               L  5e+05

Unit(concUnit="ug/L", doseUnit="mol", MW=500)

#>                Unit Factor
#> b0                   1e+00
#> CMAX           ug/L  1e+00
#> CMAXD      ug/L/mol  1e+00
#> TMAX              h  1e+00
#> TLAG              h  1e+00
#> CLST           ug/L  1e+00
#> CLSTP          ug/L  1e+00
#> TLST              h  1e+00
#> LAMZHL            h  1e+00
#> LAMZ             /h  1e+00
#> LAMZLL            h  1e+00
#> LAMZUL            h  1e+00
#> LAMZNPT              1e+00
#> CORRXY               1e+00
#> R2                   1e+00
#> R2ADJ                1e+00
#> C0             ug/L  1e+00
#> AUCLST       h*ug/L  1e+00
#> AUCALL       h*ug/L  1e+00
#> AUCIFO       h*ug/L  1e+00
#> AUCIFOD  h*ug/L/mol  1e+00
#> AUCIFP       h*ug/L  1e+00
#> AUCIFPD  h*ug/L/mol  1e+00
#> AUCPEO            %  1e+00
#> AUCPEP            %  1e+00
#> AUCPBEO           %  1e+00
#> AUCPBEP           %  1e+00
#> AUMCLST     h2*ug/L  1e+00
#> AUMCIFO     h2*ug/L  1e+00
#> AUMCIFP     h2*ug/L  1e+00
#> AUMCPEO           %  1e+00
#> AUMCPEP           %  1e+00
#> MRTIVLST          h  1e+00
#> MRTIVIFO          h  1e+00
#> MRTIVIFP          h  1e+00
#> MRTEVLST          h  1e+00
#> MRTEVIFO          h  1e+00
#> MRTEVIFP          h  1e+00
#> VZO               L  5e+08
#> VZP               L  5e+08
#> VZFO              L  5e+08
#> VZFP              L  5e+08
#> CLO             L/h  5e+08
#> CLP             L/h  5e+08
#> CLFO            L/h  5e+08
#> CLFP            L/h  5e+08
#> VSSO              L  5e+08
#> VSSP              L  5e+08

Unit(concUnit="ng/L", doseUnit="mmol", MW=500)

#>                 Unit Factor
#> b0                    1e+00
#> CMAX            ng/L  1e+00
#> CMAXD      ng/L/mmol  1e+00
#> TMAX               h  1e+00
#> TLAG               h  1e+00
#> CLST            ng/L  1e+00
#> CLSTP           ng/L  1e+00
#> TLST               h  1e+00
#> LAMZHL             h  1e+00
#> LAMZ              /h  1e+00
#> LAMZLL             h  1e+00
#> LAMZUL             h  1e+00
#> LAMZNPT               1e+00
#> CORRXY                1e+00
#> R2                    1e+00
#> R2ADJ                 1e+00
#> C0              ng/L  1e+00
#> AUCLST        h*ng/L  1e+00
#> AUCALL        h*ng/L  1e+00
#> AUCIFO        h*ng/L  1e+00
#> AUCIFOD  h*ng/L/mmol  1e+00
#> AUCIFP        h*ng/L  1e+00
#> AUCIFPD  h*ng/L/mmol  1e+00
#> AUCPEO             %  1e+00
#> AUCPEP             %  1e+00
#> AUCPBEO            %  1e+00
#> AUCPBEP            %  1e+00
#> AUMCLST      h2*ng/L  1e+00
#> AUMCIFO      h2*ng/L  1e+00
#> AUMCIFP      h2*ng/L  1e+00
#> AUMCPEO            %  1e+00
#> AUMCPEP            %  1e+00
#> MRTIVLST           h  1e+00
#> MRTIVIFO           h  1e+00
#> MRTIVIFP           h  1e+00
#> MRTEVLST           h  1e+00
#> MRTEVIFO           h  1e+00
#> MRTEVIFP           h  1e+00
#> VZO                L  5e+08
#> VZP                L  5e+08
#> VZFO               L  5e+08
#> VZFP               L  5e+08
#> CLO              L/h  5e+08
#> CLP              L/h  5e+08
#> CLFO             L/h  5e+08
#> CLFP             L/h  5e+08
#> VSSO               L  5e+08
#> VSSP               L  5e+08

Unit(concUnit="ng/mL", doseUnit="mmol", MW=500)

#>                 Unit Factor
#> b0                    1e+00
#> CMAX            ug/L  1e+00
#> CMAXD      ug/L/mmol  1e+00
#> TMAX               h  1e+00
#> TLAG               h  1e+00
#> CLST            ug/L  1e+00
#> CLSTP           ug/L  1e+00
#> TLST               h  1e+00
#> LAMZHL             h  1e+00
#> LAMZ              /h  1e+00
#> LAMZLL             h  1e+00
#> LAMZUL             h  1e+00
#> LAMZNPT               1e+00
#> CORRXY                1e+00
#> R2                    1e+00
#> R2ADJ                 1e+00
#> C0              ug/L  1e+00
#> AUCLST        h*ug/L  1e+00
#> AUCALL        h*ug/L  1e+00
#> AUCIFO        h*ug/L  1e+00
#> AUCIFOD  h*ug/L/mmol  1e+00
#> AUCIFP        h*ug/L  1e+00
#> AUCIFPD  h*ug/L/mmol  1e+00
#> AUCPEO             %  1e+00
#> AUCPEP             %  1e+00
#> AUCPBEO            %  1e+00
#> AUCPBEP            %  1e+00
#> AUMCLST      h2*ug/L  1e+00
#> AUMCIFO      h2*ug/L  1e+00
#> AUMCIFP      h2*ug/L  1e+00
#> AUMCPEO            %  1e+00
#> AUMCPEP            %  1e+00
#> MRTIVLST           h  1e+00
#> MRTIVIFO           h  1e+00
#> MRTIVIFP           h  1e+00
#> MRTEVLST           h  1e+00
#> MRTEVIFO           h  1e+00
#> MRTEVIFP           h  1e+00
#> VZO                L  5e+05
#> VZP                L  5e+05
#> VZFO               L  5e+05
#> VZFP               L  5e+05
#> CLO              L/h  5e+05
#> CLP              L/h  5e+05
#> CLFO             L/h  5e+05
#> CLFP             L/h  5e+05
#> VSSO               L  5e+05
#> VSSP               L  5e+05

Unit(concUnit="nmol/L", doseUnit="mg")

#> Warning: Molecular weight should be given for more informative results!

#>                 Unit Factor
#> b0                        1
#> CMAX          nmol/L      1
#> CMAXD      nmol/L/mg      1
#> TMAX               h      1
#> TLAG               h      1
#> CLST          nmol/L      1
#> CLSTP         nmol/L      1
#> TLST               h      1
#> LAMZHL             h      1
#> LAMZ              /h      1
#> LAMZLL             h      1
#> LAMZUL             h      1
#> LAMZNPT                   1
#> CORRXY                    1
#> R2                        1
#> R2ADJ                     1
#> C0            nmol/L      1
#> AUCLST      h*nmol/L      1
#> AUCALL      h*nmol/L      1
#> AUCIFO      h*nmol/L      1
#> AUCIFOD  h*nmol/L/mg      1
#> AUCIFP      h*nmol/L      1
#> AUCIFPD  h*nmol/L/mg      1
#> AUCPEO             %      1
#> AUCPEP             %      1
#> AUCPBEO            %      1
#> AUCPBEP            %      1
#> AUMCLST    h2*nmol/L      1
#> AUMCIFO    h2*nmol/L      1
#> AUMCIFP    h2*nmol/L      1
#> AUMCPEO            %      1
#> AUMCPEP            %      1
#> MRTIVLST           h      1
#> MRTIVIFO           h      1
#> MRTIVIFP           h      1
#> MRTEVLST           h      1
#> MRTEVIFO           h      1
#> MRTEVIFP           h      1
#> VZO                L     NA
#> VZP                L     NA
#> VZFO               L     NA
#> VZFP               L     NA
#> CLO              L/h     NA
#> CLP              L/h     NA
#> CLFO             L/h     NA
#> CLFP             L/h     NA
#> VSSO               L     NA
#> VSSP               L     NA

Unit(concUnit="ug/L", doseUnit="mmol")

#> Warning: Molecular weight should be given for more informative results!

#>                 Unit Factor
#> b0                        1
#> CMAX            ug/L      1
#> CMAXD      ug/L/mmol      1
#> TMAX               h      1
#> TLAG               h      1
#> CLST            ug/L      1
#> CLSTP           ug/L      1
#> TLST               h      1
#> LAMZHL             h      1
#> LAMZ              /h      1
#> LAMZLL             h      1
#> LAMZUL             h      1
#> LAMZNPT                   1
#> CORRXY                    1
#> R2                        1
#> R2ADJ                     1
#> C0              ug/L      1
#> AUCLST        h*ug/L      1
#> AUCALL        h*ug/L      1
#> AUCIFO        h*ug/L      1
#> AUCIFOD  h*ug/L/mmol      1
#> AUCIFP        h*ug/L      1
#> AUCIFPD  h*ug/L/mmol      1
#> AUCPEO             %      1
#> AUCPEP             %      1
#> AUCPBEO            %      1
#> AUCPBEP            %      1
#> AUMCLST      h2*ug/L      1
#> AUMCIFO      h2*ug/L      1
#> AUMCIFP      h2*ug/L      1
#> AUMCPEO            %      1
#> AUMCPEP            %      1
#> MRTIVLST           h      1
#> MRTIVIFO           h      1
#> MRTIVIFP           h      1
#> MRTEVLST           h      1
#> MRTEVIFO           h      1
#> MRTEVIFP           h      1
#> VZO                L     NA
#> VZP                L     NA
#> VZFO               L     NA
#> VZFP               L     NA
#> CLO              L/h     NA
#> CLP              L/h     NA
#> CLFO             L/h     NA
#> CLFP             L/h     NA
#> VSSO               L     NA
#> VSSP               L     NA