Noncompartmental Analysis for an Individual

It performs a noncompartmental analysis with one subject data

IndiNCA(x, y, dose = 0, fit = "Linear", adm = "Extravascular", dur = 0,
               report = "Table", iAUC = "", uTime = "h", uConc = "ug/L", uDose = "mg")

Arguments

x
vector values of independent variable, usually time

y
vector values of dependent variable, usually concentration

dose
administered dose for the subject

fit
either of "Linear" or "Log" to indicate the way to calculate AUC and AUMC

adm
one of "Bolus" or "Infusion" or "Extravascular" to indicate drug administration mode

dur
infusion duration for constant infusion, otherwise 0

report
either of "Table" or "Text" to specify the type of return value

iAUC
data.frame with three columns, "Name", "Start", "End" to specify the invervals for partial (interval) AUC

uTime
unit of time

uConc
unit of concentration

uDose
unit of dose

Details

This performs a noncompartmental analysis for a subject. It returns practically the same result with the most popular commercial software.

Value

CMAX

maximum concentration, Cmax

CMAXD

dose normalized Cmax, CMAX / Dose, Cmax / Dose

TMAX

time of maximum concentration, Tmax

TLAG

time to observe the first non-zero concentration, for extravascular administration only

CLST

last positive concentration observed, Clast

CLSTP

last positive concentration predicted, Clast_pred

TLST

time of last positive concentration, Tlast

LAMZHL

half-life by lambda z, ln(2)/LAMZ

LAMZ

lambda_z negative of best fit terminal slope

LAMZLL

earliest time for LAMZ

LAMZUL

last time for LAMZ

LAMZNPT

number of points for LAMZ

CORRXY

correlation of log(concentration) and time

R2

R-squared

R2ADJ

R-squared adjusted

C0

back extrapolated concentration at time 0, for bolus intravascular administration only

AUCLST

AUC from 0 to TLST

AUCALL

AUC using all the given points, including trailing zero concentrations

AUCIFO

AUC infinity observed

AUCIFOD

AUCIFO / Dose

AUCIFP

AUC infinity predicted using CLSTP instead of CLST

AUCIFPD

AUCIFP / Dose

AUCPEO

AUC % extrapolation observed

AUCPEP

AUC % extrapolated for AUCIFP

AUCPBEO

AUC % back extrapolation observed, for bolus IV administration only

AUCPBEP

AUC % back extrapolation predicted with AUCIFP, for bolus IV administration only

AUMCLST

AUMC to the TLST

AUMCIFO

AUMC infinity observed using CLST

AUMCIFP

AUMC infinity determined by CLSTP

AUMCPEO

AUMC % extrapolated observed

AUMCPEP

AUMC % extrapolated predicted

MRTIVLST

mean residence time (MRT) to TLST, for intravascular administration

MRTIVIFO

mean residence time (MRT) infinity using CLST, for intravascular administration

MRTIVIFP

mean residence time (MRT) infinity using CLSTP, for intravascular administration

MRTEVLST

mean residence time (MRT) to TLST, for extravascular administration

MRTEVIFO

mean residence time (MRT) infinity using CLST, for extravascular administration

MRTEVIFP

mean residence time (MRT) infinity using CLSTP, for extravascular administration

VZO

volume of distribution determined by LAMZ and AUCIFO, for intravascular administration

VZP

volume of distribution determined by LAMZ and AUCIFP, for intravascular administration

VZFO

VZO for extravascular administration, VZO/F, F is bioavailability

VZFP

VZP for extravascular administration, VZP/F, F is bioavailability

CLO

clearance using AUCIFO, for intravascular administration

CLP

clearance using AUCIFP, for intravascular administration

CLFO

CLO for extravascular administration, CLO/F, F is bioavailability

CLFP

CLP for extravascular administration, CLP/F, F is bioavailability

VSSO

volume of distribution at steady state using CLST, for intravascular administration only

VSSP

volume of distribution at stead state using CLSTP, for intravascular administration only

References

Gabrielsson J, Weiner D. Pharmacokinetic and Pharmacodynamic Data Analysis - Concepts and Applications. 5th ed. 2016.

Shargel L, Yu A. Applied Biopharmaceutics and Pharmacokinetics. 7th ed. 2015.

Rowland M, Tozer TN. Clinical Pharmacokinetics and Pharmacodynamics - Concepts and Applications. 4th ed. 2011.

Gibaldi M, Perrier D. Pharmacokinetics. 2nd ed. revised and expanded. 1982.

See also

AUC, BestSlope

Examples

IndiNCA(Theoph[Theoph$Subject==1,"Time"], Theoph[Theoph$Subject==1, "conc"], dose=320, uConc="mg/L")

#>         CMAX        CMAXD         TMAX         TLAG         CLST        CLSTP 
#>   10.5000000    0.0328125    1.1200000    0.0000000    3.2800000    3.2801465 
#>         TLST       LAMZHL         LAMZ       LAMZLL       LAMZUL      LAMZNPT 
#>   24.3700000   14.3043776    0.0484570    9.0500000   24.3700000    3.0000000 
#>       CORRXY           R2        R2ADJ       AUCLST       AUCALL       AUCIFO 
#>   -0.9999999    0.9999997    0.9999995  148.9230500  148.9230500  216.6119330 
#>      AUCIFOD       AUCPEO       AUCIFP      AUCIFPD       AUCPEP      AUMCLST 
#>    0.6769123   31.2489169  216.6149558    0.6769217   31.2498763 1459.0711035 
#>      AUMCIFO      AUMCPEO      AUMCIFP      AUMCPEP     MRTEVLST     MRTEVIFO 
#> 4505.5348194   67.6160287 4505.6708646   67.6170065    9.7974834   20.8000305 
#>     MRTEVIFP         VZFO         VZFP         CLFO         CLFP 
#>   20.8003683   30.4867482   30.4863228    1.4772963    1.4772757 

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Bolus", uConc="mg/L")

#>        CMAX       CMAXD        TMAX          C0     AUCPBEO     AUCPBEP 
#>  1.50000000  0.06000000  0.25000000  2.39361702 20.65564214 20.64189142 
#>        CLST       CLSTP        TLST      LAMZHL        LAMZ      LAMZLL 
#>  0.05000000  0.05024851  8.00000000  4.37812701  0.15832048  5.00000000 
#>      LAMZUL     LAMZNPT      CORRXY          R2       R2ADJ      AUCLST 
#>  8.00000000  3.00000000 -0.99853229  0.99706673  0.99413345  2.04045213 
#>      AUCALL      AUCIFO     AUCIFOD      AUCPEO      AUCIFP     AUCIFPD 
#>  2.04045213  2.35626723  0.09425069 13.40319561  2.35783688  0.09431348 
#>      AUCPEP     AUMCLST     AUMCIFO     AUMCPEO     AUMCIFP     AUMCPEP 
#> 13.46084419  3.27125000  7.79255448 58.02082606  7.81502594 58.14153371 
#>    MRTIVLST    MRTIVIFO    MRTIVIFP         VZO         VZP         CLO 
#>  1.60319860  3.30716074  3.31448966 67.01597804 66.97136466 10.61000197 
#>         CLP        VSSO        VSSP 
#> 10.60293876 35.08898193 35.14333088 

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Infusion", dur=0.25, uConc="mg/L")

#>        CMAX       CMAXD        TMAX        CLST        TLST       CLSTP 
#>  1.50000000  0.06000000  0.25000000  0.05000000  8.00000000  0.05024851 
#>      LAMZHL        LAMZ      LAMZLL      LAMZUL     LAMZNPT      CORRXY 
#>  4.37812701  0.15832048  5.00000000  8.00000000  3.00000000 -0.99853229 
#>          R2       R2ADJ      AUCLST      AUCALL      AUCIFO     AUCIFOD 
#>  0.99706673  0.99413345  1.74125000  1.74125000  2.05706511  0.08228260 
#>      AUCPEO      AUCIFP     AUCIFPD      AUCPEP     AUMCLST     AUMCIFO 
#> 15.35270349  2.05863475  0.08234539 15.41724428  3.27125000  7.79255448 
#>     AUMCPEO     AUMCIFP     AUMCPEP    MRTIVLST    MRTIVIFO    MRTIVIFP 
#> 58.02082606  7.81502594 58.14153371  1.75367911  3.66319049  3.67121783 
#>         VZO         VZP         CLO         CLP        VSSO        VSSP 
#> 76.76351746 76.70498779 12.15323712 12.14397067 44.51962266 44.58316170 

IndiNCA(Theoph[Theoph$Subject==1,"Time"], Theoph[Theoph$Subject==1, "conc"], dose=320,
        report="Text", uConc="mg/L")

#>  [1] "                        NONCOMPARTMENTAL ANALYSIS REPORT"                           
#>  [2] "                       Package version 0.1.1 (2017-07-10)"                          
#>  [3] "                          R version 3.4.2 (2017-09-28)"                             
#>  [4] ""                                                                                   
#>  [5] "Date and Time: 2017-10-20 10:38:06 Asia/Seoul"                                      
#>  [6] ""                                                                                   
#>  [7] "Calculation Setting"                                                                
#>  [8] "-------------------"                                                                
#>  [9] "Drug Administration: Extravascular"                                                 
#> [10] "Observation count excluding trailing zero: 11"                                      
#> [11] "dose at time 0: 320"                                                                
#> [12] "AUC Calculation Method: Linear-up Linear-down fit"                                  
#> [13] "Weighting for lambda z: Uniform (Ordinary Least Square, OLS)"                       
#> [14] "Lambda z selection criterion: Heighest adjusted R-squared value with precision=1e-4"
#> [15] ""                                                                                   
#> [16] ""                                                                                   
#> [17] "Fitting, AUC, AUMC Result"                                                          
#> [18] "-------------------------"                                                          
#> [19] "      Time         Conc.      Pred.   Residual       AUC       AUMC      Weight"    
#> [20] "-------------------------------------------------------------------------------"    
#> [21] "     0.0000       0.7400                           0.0000     0.0000 "              
#> [22] "     0.2500       2.8400                           0.4475     0.0888 "              
#> [23] "     0.5700       6.5700                           1.9531     0.8015 "              
#> [24] "     1.1200      10.5000                           6.6474     5.0654 "              
#> [25] "     2.0200       9.6600                          15.7194    19.1383 "              
#> [26] "     3.8200       8.5800                          32.1354    66.1982 "              
#> [27] "     5.1000       8.3600                          42.9769   114.4617 "              
#> [28] "     7.0300       7.4700                          58.2529   206.2815 "              
#> [29] "     9.0500 *     6.8900     6.8912 -1.228e-03    72.7565   322.2988     1.0000"    
#> [30] "    12.1200 *     5.9400     5.9387 +1.324e-03    92.4505   528.5219     1.0000"    
#> [31] "    24.3700 *     3.2800     3.2801 -1.465e-04   148.9231  1459.0711     1.0000"    
#> [32] ""                                                                                   
#> [33] "*: Used for the calculation of Lambda z."                                           
#> [34] ""                                                                                   
#> [35] ""                                                                                   
#> [36] "Calculated Values"                                                                  
#> [37] "-----------------"                                                                  
#> [38] "CMAX       Max Conc                                       10.5000 mg/L"             
#> [39] "CMAXD      Max Conc Norm by Dose                           0.0328 mg/L/mg"          
#> [40] "TMAX       Time of CMAX                                    1.1200 h"                
#> [41] "TLAG       Time Until First Nonzero Conc                   0.0000 h"                
#> [42] "CLST       Last Nonzero Conc                               3.2800 mg/L"             
#> [43] "CLSTP      Last Nonzero Conc Pred                          3.2801 mg/L"             
#> [44] "TLST       Time of Last Nonzero Conc                      24.3700 h"                
#> [45] "LAMZHL     Half-Life Lambda z                             14.3044 h"                
#> [46] "LAMZ       Lambda z                                        0.0485 /h"               
#> [47] "LAMZLL     Lambda z Lower Limit                            9.0500 h"                
#> [48] "LAMZUL     Lambda z Upper Limit                           24.3700 h"                
#> [49] "LAMZNPT    Number of Points for Lambda z                   3"                       
#> [50] "CORRXY     Correlation Between TimeX and Log ConcY        -1.0000 "                 
#> [51] "R2         R Squared                                       1.0000 "                 
#> [52] "R2ADJ      R Squared Adjusted                              1.0000 "                 
#> [53] "AUCLST     AUC to Last Nonzero Conc                      148.9231 h*mg/L"           
#> [54] "AUCALL     AUC All                                       148.9231 h*mg/L"           
#> [55] "AUCIFO     AUC Infinity Obs                              216.6119 h*mg/L"           
#> [56] "AUCIFOD    AUC Infinity Obs Norm by Dose                   0.6769 h*mg/L/mg"        
#> [57] "AUCPEO     AUC %Extrapolation Obs                         31.2489 %"                
#> [58] "AUCIFP     AUC Infinity Pred                             216.6150 h*mg/L"           
#> [59] "AUCIFPD    AUC Infinity Pred Norm by Dose                  0.6769 h*mg/L/mg"        
#> [60] "AUCPEP     AUC %Extrapolation Pred                        31.2499 %"                
#> [61] "AUMCLST    AUMC to Last Nonzero Conc                    1459.0711 h2*mg/L"          
#> [62] "AUMCIFO    AUMC Infinity Obs                            4505.5348 h2*mg/L"          
#> [63] "AUMCPEO    AUMC %Extrapolation Obs                        67.6160 %"                
#> [64] "AUMCIFP    AUMC Infinity Pred                           4505.6709 h2*mg/L"          
#> [65] "AUMCPEP    AUMC % Extrapolation Pred                      67.6170 %"                
#> [66] "MRTEVLST   MRT Extravasc to Last Nonzero Conc              9.7975 h"                
#> [67] "MRTEVIFO   MRT Extravasc Infinity Obs                     20.8000 h"                
#> [68] "MRTEVIFP   MRT Extravasc Infinity Pred                    20.8004 h"                
#> [69] "VZFO       Vz Obs by F                                    30.4867 L"                
#> [70] "VZFP       Vz Pred by F                                   30.4863 L"                
#> [71] "CLFO       Total CL Obs by F                               1.4773 L/h"              
#> [72] "CLFP       Total CL Pred by F                              1.4773 L/h"              

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Bolus", report="Text", uConc="mg/L")

#>  [1] "                        NONCOMPARTMENTAL ANALYSIS REPORT"                           
#>  [2] "                       Package version 0.1.1 (2017-07-10)"                          
#>  [3] "                          R version 3.4.2 (2017-09-28)"                             
#>  [4] ""                                                                                   
#>  [5] "Date and Time: 2017-10-20 10:38:06 Asia/Seoul"                                      
#>  [6] ""                                                                                   
#>  [7] "Calculation Setting"                                                                
#>  [8] "-------------------"                                                                
#>  [9] "Drug Administration: Bolus IV"                                                      
#> [10] "Observation count excluding trailing zero: 11"                                      
#> [11] "dose at time 0: 25"                                                                 
#> [12] "AUC Calculation Method: Linear-up Linear-down fit"                                  
#> [13] "Weighting for lambda z: Uniform (Ordinary Least Square, OLS)"                       
#> [14] "Lambda z selection criterion: Heighest adjusted R-squared value with precision=1e-4"
#> [15] ""                                                                                   
#> [16] ""                                                                                   
#> [17] "Fitting, AUC, AUMC Result"                                                          
#> [18] "-------------------------"                                                          
#> [19] "      Time         Conc.      Pred.   Residual       AUC       AUMC      Weight"    
#> [20] "-------------------------------------------------------------------------------"    
#> [21] "     0.0000 +     2.3936                           0.0000     0.0000 "              
#> [22] "     0.2500       1.5000                           0.4867     0.0469 "              
#> [23] "     0.5000       0.9400                           0.7917     0.1525 "              
#> [24] "     0.7500       0.7800                           1.0067     0.2844 "              
#> [25] "     1.0000       0.4800                           1.1642     0.4175 "              
#> [26] "     1.2500       0.3700                           1.2705     0.5353 "              
#> [27] "     2.0000       0.1900                           1.4805     0.8513 "              
#> [28] "     3.0000       0.1200                           1.6355     1.2213 "              
#> [29] "     4.0000       0.1100                           1.7505     1.6213 "              
#> [30] "     5.0000 *     0.0800     0.0808 -7.972e-04     1.8455     2.0413     1.0000"    
#> [31] "     6.0000 *     0.0700     0.0690 +1.033e-03     1.9205     2.4513     1.0000"    
#> [32] "     8.0000 *     0.0500     0.0502 -2.485e-04     2.0405     3.2713     1.0000"    
#> [33] ""                                                                                   
#> [34] "+: Back extrapolated concentration"                                                 
#> [35] "*: Used for the calculation of Lambda z."                                           
#> [36] ""                                                                                   
#> [37] ""                                                                                   
#> [38] "Calculated Values"                                                                  
#> [39] "-----------------"                                                                  
#> [40] "CMAX       Max Conc                                        1.5000 mg/L"             
#> [41] "CMAXD      Max Conc Norm by Dose                           0.0600 mg/L/mg"          
#> [42] "TMAX       Time of CMAX                                    0.2500 h"                
#> [43] "C0         Initial Conc                                    2.3936 mg/L"             
#> [44] "AUCPBEO    AUC %Back Extrapolation Obs                    20.6556 %"                
#> [45] "AUCPBEP    AUC %Back Extrapolation Pred                   20.6419 %"                
#> [46] "CLST       Last Nonzero Conc                               0.0500 mg/L"             
#> [47] "CLSTP      Last Nonzero Conc Pred                          0.0502 mg/L"             
#> [48] "TLST       Time of Last Nonzero Conc                       8.0000 h"                
#> [49] "LAMZHL     Half-Life Lambda z                              4.3781 h"                
#> [50] "LAMZ       Lambda z                                        0.1583 /h"               
#> [51] "LAMZLL     Lambda z Lower Limit                            5.0000 h"                
#> [52] "LAMZUL     Lambda z Upper Limit                            8.0000 h"                
#> [53] "LAMZNPT    Number of Points for Lambda z                   3"                       
#> [54] "CORRXY     Correlation Between TimeX and Log ConcY        -0.9985 "                 
#> [55] "R2         R Squared                                       0.9971 "                 
#> [56] "R2ADJ      R Squared Adjusted                              0.9941 "                 
#> [57] "AUCLST     AUC to Last Nonzero Conc                        2.0405 h*mg/L"           
#> [58] "AUCALL     AUC All                                         2.0405 h*mg/L"           
#> [59] "AUCIFO     AUC Infinity Obs                                2.3563 h*mg/L"           
#> [60] "AUCIFOD    AUC Infinity Obs Norm by Dose                   0.0943 h*mg/L/mg"        
#> [61] "AUCPEO     AUC %Extrapolation Obs                         13.4032 %"                
#> [62] "AUCIFP     AUC Infinity Pred                               2.3578 h*mg/L"           
#> [63] "AUCIFPD    AUC Infinity Pred Norm by Dose                  0.0943 h*mg/L/mg"        
#> [64] "AUCPEP     AUC %Extrapolation Pred                        13.4608 %"                
#> [65] "AUMCLST    AUMC to Last Nonzero Conc                       3.2713 h2*mg/L"          
#> [66] "AUMCIFO    AUMC Infinity Obs                               7.7926 h2*mg/L"          
#> [67] "AUMCPEO    AUMC %Extrapolation Obs                        58.0208 %"                
#> [68] "AUMCIFP    AUMC Infinity Pred                              7.8150 h2*mg/L"          
#> [69] "AUMCPEP    AUMC % Extrapolation Pred                      58.1415 %"                
#> [70] "MRTIVLST   MRT Intravasc to Last Nonzero Conc              1.6032 h"                
#> [71] "MRTIVIFO   MRT Intravasc Infinity Obs                      3.3072 h"                
#> [72] "MRTIVIFP   MRT Intravasc Infinity Pred                     3.3145 h"                
#> [73] "VZO        Vz Obs                                         67.0160 L"                
#> [74] "VZP        Vz Pred                                        66.9714 L"                
#> [75] "CLO        Total CL Obs                                   10.6100 L/h"              
#> [76] "CLP        Total CL Pred                                  10.6029 L/h"              
#> [77] "VSSO       Vol Dist Steady State Obs                      35.0890 L"                
#> [78] "VSSP       Vol Dist Steady State Pred                     35.1433 L"                

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Infusion", dur=0.25, report="Text", uConc="mg/L")

#>  [1] "                        NONCOMPARTMENTAL ANALYSIS REPORT"                           
#>  [2] "                       Package version 0.1.1 (2017-07-10)"                          
#>  [3] "                          R version 3.4.2 (2017-09-28)"                             
#>  [4] ""                                                                                   
#>  [5] "Date and Time: 2017-10-20 10:38:06 Asia/Seoul"                                      
#>  [6] ""                                                                                   
#>  [7] "Calculation Setting"                                                                
#>  [8] "-------------------"                                                                
#>  [9] "Drug Administration: Constant Infusion"                                             
#> [10] "Observation count excluding trailing zero: 11"                                      
#> [11] "dose at time 0: 25"                                                                 
#> [12] "Length of Infusion: 0.25"                                                           
#> [13] "AUC Calculation Method: Linear-up Linear-down fit"                                  
#> [14] "Weighting for lambda z: Uniform (Ordinary Least Square, OLS)"                       
#> [15] "Lambda z selection criterion: Heighest adjusted R-squared value with precision=1e-4"
#> [16] ""                                                                                   
#> [17] ""                                                                                   
#> [18] "Fitting, AUC, AUMC Result"                                                          
#> [19] "-------------------------"                                                          
#> [20] "      Time         Conc.      Pred.   Residual       AUC       AUMC      Weight"    
#> [21] "-------------------------------------------------------------------------------"    
#> [22] "     0.0000 +     0.0000                           0.0000     0.0000 "              
#> [23] "     0.2500       1.5000                           0.1875     0.0469 "              
#> [24] "     0.5000       0.9400                           0.4925     0.1525 "              
#> [25] "     0.7500       0.7800                           0.7075     0.2844 "              
#> [26] "     1.0000       0.4800                           0.8650     0.4175 "              
#> [27] "     1.2500       0.3700                           0.9713     0.5353 "              
#> [28] "     2.0000       0.1900                           1.1813     0.8513 "              
#> [29] "     3.0000       0.1200                           1.3363     1.2213 "              
#> [30] "     4.0000       0.1100                           1.4513     1.6213 "              
#> [31] "     5.0000 *     0.0800     0.0808 -7.972e-04     1.5462     2.0413     1.0000"    
#> [32] "     6.0000 *     0.0700     0.0690 +1.033e-03     1.6213     2.4513     1.0000"    
#> [33] "     8.0000 *     0.0500     0.0502 -2.485e-04     1.7413     3.2713     1.0000"    
#> [34] ""                                                                                   
#> [35] "+: Back extrapolated concentration"                                                 
#> [36] "*: Used for the calculation of Lambda z."                                           
#> [37] ""                                                                                   
#> [38] ""                                                                                   
#> [39] "Calculated Values"                                                                  
#> [40] "-----------------"                                                                  
#> [41] "CMAX       Max Conc                                        1.5000 mg/L"             
#> [42] "CMAXD      Max Conc Norm by Dose                           0.0600 mg/L/mg"          
#> [43] "TMAX       Time of CMAX                                    0.2500 h"                
#> [44] "CLST       Last Nonzero Conc                               0.0500 mg/L"             
#> [45] "TLST       Time of Last Nonzero Conc                       8.0000 h"                
#> [46] "CLSTP      Last Nonzero Conc Pred                          0.0502 mg/L"             
#> [47] "LAMZHL     Half-Life Lambda z                              4.3781 h"                
#> [48] "LAMZ       Lambda z                                        0.1583 /h"               
#> [49] "LAMZLL     Lambda z Lower Limit                            5.0000 h"                
#> [50] "LAMZUL     Lambda z Upper Limit                            8.0000 h"                
#> [51] "LAMZNPT    Number of Points for Lambda z                   3"                       
#> [52] "CORRXY     Correlation Between TimeX and Log ConcY        -0.9985 "                 
#> [53] "R2         R Squared                                       0.9971 "                 
#> [54] "R2ADJ      R Squared Adjusted                              0.9941 "                 
#> [55] "AUCLST     AUC to Last Nonzero Conc                        1.7413 h*mg/L"           
#> [56] "AUCALL     AUC All                                         1.7413 h*mg/L"           
#> [57] "AUCIFO     AUC Infinity Obs                                2.0571 h*mg/L"           
#> [58] "AUCIFOD    AUC Infinity Obs Norm by Dose                   0.0823 h*mg/L/mg"        
#> [59] "AUCPEO     AUC %Extrapolation Obs                         15.3527 %"                
#> [60] "AUCIFP     AUC Infinity Pred                               2.0586 h*mg/L"           
#> [61] "AUCIFPD    AUC Infinity Pred Norm by Dose                  0.0823 h*mg/L/mg"        
#> [62] "AUCPEP     AUC %Extrapolation Pred                        15.4172 %"                
#> [63] "AUMCLST    AUMC to Last Nonzero Conc                       3.2713 h2*mg/L"          
#> [64] "AUMCIFO    AUMC Infinity Obs                               7.7926 h2*mg/L"          
#> [65] "AUMCPEO    AUMC %Extrapolation Obs                        58.0208 %"                
#> [66] "AUMCIFP    AUMC Infinity Pred                              7.8150 h2*mg/L"          
#> [67] "AUMCPEP    AUMC % Extrapolation Pred                      58.1415 %"                
#> [68] "MRTIVLST   MRT Intravasc to Last Nonzero Conc              1.7537 h"                
#> [69] "MRTIVIFO   MRT Intravasc Infinity Obs                      3.6632 h"                
#> [70] "MRTIVIFP   MRT Intravasc Infinity Pred                     3.6712 h"                
#> [71] "VZO        Vz Obs                                         76.7635 L"                
#> [72] "VZP        Vz Pred                                        76.7050 L"                
#> [73] "CLO        Total CL Obs                                   12.1532 L/h"              
#> [74] "CLP        Total CL Pred                                  12.1440 L/h"              
#> [75] "VSSO       Vol Dist Steady State Obs                      44.5196 L"                
#> [76] "VSSP       Vol Dist Steady State Pred                     44.5832 L"                

iAUC = data.frame(Name=c("AUC[0-12h]","AUC[0-24h]"), Start=c(0,0), End=c(12,24)) ; iAUC

#>         Name Start End
#> 1 AUC[0-12h]     0  12
#> 2 AUC[0-24h]     0  24

IndiNCA(Theoph[Theoph$Subject==1,"Time"], Theoph[Theoph$Subject==1, "conc"], dose=320,
        iAUC=iAUC, uConc="mg/L")

#> Warning: length > 1 ¿Ã∂Û¥¬ ¡∂∞«¿Ã ¿÷∞Ì, √ππ¯¬∞ ø‰º“∏∏¿Ã ªÁøÎµ… ∞Õ¿‘¥œ¥Ÿ

#> Warning: length > 1 ¿Ã∂Û¥¬ ¡∂∞«¿Ã ¿÷∞Ì, √ππ¯¬∞ ø‰º“∏∏¿Ã ªÁøÎµ… ∞Õ¿‘¥œ¥Ÿ

#>         CMAX        CMAXD         TMAX         TLAG         CLST        CLSTP 
#>   10.5000000    0.0328125    1.1200000    0.0000000    3.2800000    3.2801465 
#>         TLST       LAMZHL         LAMZ       LAMZLL       LAMZUL      LAMZNPT 
#>   24.3700000   14.3043776    0.0484570    9.0500000   24.3700000    3.0000000 
#>       CORRXY           R2        R2ADJ       AUCLST       AUCALL       AUCIFO 
#>   -0.9999999    0.9999997    0.9999995  148.9230500  148.9230500  216.6119330 
#>      AUCIFOD       AUCPEO       AUCIFP      AUCIFPD       AUCPEP      AUMCLST 
#>    0.6769123   31.2489169  216.6149558    0.6769217   31.2498763 1459.0711035 
#>      AUMCIFO      AUMCPEO      AUMCIFP      AUMCPEP     MRTEVLST     MRTEVIFO 
#> 4505.5348194   67.6160287 4505.6708646   67.6170065    9.7974834   20.8000305 
#>     MRTEVIFP         VZFO         VZFP         CLFO         CLFP   AUC[0-12h] 
#>   20.8003683   30.4867482   30.4863228    1.4772963    1.4772757   91.7355220 
#>   AUC[0-24h] 
#>  147.6945866 

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Bolus", iAUC=iAUC, uConc="mg/L")

#> Warning: length > 1 ¿Ã∂Û¥¬ ¡∂∞«¿Ã ¿÷∞Ì, √ππ¯¬∞ ø‰º“∏∏¿Ã ªÁøÎµ… ∞Õ¿‘¥œ¥Ÿ

#> Warning: length > 1 ¿Ã∂Û¥¬ ¡∂∞«¿Ã ¿÷∞Ì, √ππ¯¬∞ ø‰º“∏∏¿Ã ªÁøÎµ… ∞Õ¿‘¥œ¥Ÿ

#>        CMAX       CMAXD        TMAX          C0     AUCPBEO     AUCPBEP 
#>  1.50000000  0.06000000  0.25000000  2.39361702 20.65564214 20.64189142 
#>        CLST       CLSTP        TLST      LAMZHL        LAMZ      LAMZLL 
#>  0.05000000  0.05024851  8.00000000  4.37812701  0.15832048  5.00000000 
#>      LAMZUL     LAMZNPT      CORRXY          R2       R2ADJ      AUCLST 
#>  8.00000000  3.00000000 -0.99853229  0.99706673  0.99413345  2.04045213 
#>      AUCALL      AUCIFO     AUCIFOD      AUCPEO      AUCIFP     AUCIFPD 
#>  2.04045213  2.35626723  0.09425069 13.40319561  2.35783688  0.09431348 
#>      AUCPEP     AUMCLST     AUMCIFO     AUMCPEO     AUMCIFP     AUMCPEP 
#> 13.46084419  3.27125000  7.79255448 58.02082606  7.81502594 58.14153371 
#>    MRTIVLST    MRTIVIFO    MRTIVIFP         VZO         VZP         CLO 
#>  1.60319860  3.30716074  3.31448966 67.01597804 66.97136466 10.61000197 
#>         CLP        VSSO        VSSP  AUC[0-12h]  AUC[0-24h] 
#> 10.60293876 35.08898193 35.14333088  2.18894711  2.33163353 

IndiNCA(Indometh[Indometh$Subject==1,"time"], Indometh[Indometh$Subject==1, "conc"], dose=25,
        adm="Infusion", dur=0.25, iAUC=iAUC, uConc="mg/L")

#> Warning: length > 1 ¿Ã∂Û¥¬ ¡∂∞«¿Ã ¿÷∞Ì, √ππ¯¬∞ ø‰º“∏∏¿Ã ªÁøÎµ… ∞Õ¿‘¥œ¥Ÿ

#> Warning: length > 1 ¿Ã∂Û¥¬ ¡∂∞«¿Ã ¿÷∞Ì, √ππ¯¬∞ ø‰º“∏∏¿Ã ªÁøÎµ… ∞Õ¿‘¥œ¥Ÿ

#>        CMAX       CMAXD        TMAX        CLST        TLST       CLSTP 
#>  1.50000000  0.06000000  0.25000000  0.05000000  8.00000000  0.05024851 
#>      LAMZHL        LAMZ      LAMZLL      LAMZUL     LAMZNPT      CORRXY 
#>  4.37812701  0.15832048  5.00000000  8.00000000  3.00000000 -0.99853229 
#>          R2       R2ADJ      AUCLST      AUCALL      AUCIFO     AUCIFOD 
#>  0.99706673  0.99413345  1.74125000  1.74125000  2.05706511  0.08228260 
#>      AUCPEO      AUCIFP     AUCIFPD      AUCPEP     AUMCLST     AUMCIFO 
#> 15.35270349  2.05863475  0.08234539 15.41724428  3.27125000  7.79255448 
#>     AUMCPEO     AUMCIFP     AUMCPEP    MRTIVLST    MRTIVIFO    MRTIVIFP 
#> 58.02082606  7.81502594 58.14153371  1.75367911  3.66319049  3.67121783 
#>         VZO         VZP         CLO         CLP        VSSO        VSSP 
#> 76.76351746 76.70498779 12.15323712 12.14397067 44.51962266 44.58316170 
#>  AUC[0-12h]  AUC[0-24h] 
#>  1.88974498  2.03243140