Plot concentration vs. time curve for individuals and collectively.

Generates individual and superposed concentration vs. time curve and save it in pdf files.

plotPK(concData, id, Time, conc, unitTime = "hr", unitConc = "ng/mL", trt = "",
       fit = "Linear", dose = 0, adm = "Extravascular", dur = 0, outdir = "Output")

Arguments

concData
name of data table containing time-concentration data of multiple subjects

id
column name for subject ID

Time
column name for the time

conc
column name for the concentration

unitTime
unit for the time

unitConc
unit for the concentration

trt
column name for the treatment code. This is useful for crossover study like bioequivalence trial.

fit
one of "Linear" or "Log" to indicate the way to calculate AUC

dose
administered dose. One should be careful for the unit. This can be a vector containing dose for each subject in order.

adm
one of "Bolus" or "Infusion" or "Extravascular" to indicate drug administration mode

dur
infusion duration for constant infusion, otherwise 0. This can be a vector containing values for each subject in order.

outdir
name of the folder to be used for the output files

Details

This function generates plots for individual and summary concentration vs. time curve. This function calles NCA().

Value

This function saves pdf files and tiff files in the outdir folder.

See also

NCA

Examples

plotPK(Theoph, "Subject", "Time", "conc", unitTime="hr", unitConc="mg/L", dose=320)

#> pdf 
#>   2 

plotPK(Indometh, "Subject", "time", "conc", unitTime="hr", unitConc="mg/L", adm="Bolus", dose=25)

#> pdf 
#>   2