Do NCA for review

This performs NCA from the CDISC EX and PC datasets.

rNCA(ex, pc, study = "", trt = "", id = "", analyte = "",
     codeBQL = c("< 0", "<0", "NQ", "BLQ", "BQL", "BQoL", "<LOQ"),
     fit="Linear", MinPoints = 5)

Arguments

ex
EX domain data, usually from the loadEXPC

pc
PC domain data, usually form the loadEXPC

study
vector of study names in EX and PC domain to do NCA

trt
vector of treatment names in EXTRT to do NCA

id
vector of subject IDs in USUBJID to do NCA

analyte
vector of molecular species in PCTESTCD to do NCA

codeBQL
symbols of below the quantitation limit

fit
either of "Linear" or "Log" to indicate the way to calculate AUC and AUMC

MinPoints
minimum number of sampling points for NCA

Details

This calls NCA0. Results of this can be further processed by foreNCA to plot and compare between studies and dose groups.

Value

This returns a table of NCA results

See also